### 2023-11-09 #6044 @editorialbot
[PRE REVIEW]: ETHOS.PeNALPS: A Tool for the Load Profile Simulation of Industrial Processes Based on a Material Flow Simulation
**Submitting author:** @julian-belina (<a href="http://orcid.org/0000-0002-5878-2936">Julian Belina</a>)
**Repository:** https://github.com/FZJ-IEK3-VSA/ETHOS_PeNALPS
**Branch with paper.md** (empty if default branch):
**Version:** 1.0.0
