### 2023-11-09 #6044 @kyleniemeyer
Hi @julian-belina, thanks for your submission to JOSS. Unfortunately, we do not have an editor available to handle this at the moment, so I have to put your paper on our waitlist until someone frees up.

In the meantime, can you review the DOI warnings/errors for the paper? Also, any reviewer suggestions you can make would be welcome.
