+++ src/ethos_penalps/post_processing/report_generator/process_chain_report_generator.py
@@ -16,7 +16,7 @@ from reportlab.graphics import renderPM
 from svglib.svglib import svg2rlg
 
 from ethos_penalps.data_classes import CurrentProcessNode, LoopCounter
-from ethos_penalps.debugging_information import (
+from ethos_penalps.utilities.debugging_information import (
     DebuggingInformationLogger,
     NodeOperationViewer,
 )
