### 2024-02-08 #6044 @jamesdamillington
Sorry this is outside my expertise and I am already (slowly!) reviewing another submission currently for JOSS. So I will have to pass on this one.
