### 2024-02-16 #6358 @srmnitc
@stoldjo Thanks for agreeing to review.  Here is some information to get started:

> 👋🏼 @julian-belina @dpzhuX @Extraweich this is the review thread for the paper. All of our communications will happen here from now on.
>
> As a reviewer, the first step is to create a checklist for your review by entering
>
> `@editorialbot generate my checklist`
>
> as the top of a new comment in this thread.
>
> These checklists contain the JOSS requirements. As you go over the submission, please check any items that you feel have been satisfied. The first comment in this thread also contains links to the JOSS reviewer guidelines.
>
> The JOSS review is different from most other journals. Our goal is to work with the authors to help them meet our criteria instead of merely passing judgment on the submission. As such, the reviewers are encouraged to submit issues and pull requests on the software repository. When doing so, please mention `openjournals/joss-reviews#REVIEW_NUMBER` so that a link is created to this thread (and I can keep an eye on what is happening). Please also feel free to comment and ask questions on this thread. In my experience, it is better to post comments/questions/suggestions as you come across them instead of waiting until you've reviewed the entire package.
>
> We aim for reviews to be completed within about 2-4 weeks. Please let me know if any of you require some more time. We can also use EditorialBot (our bot) to set automatic reminders if you know you'll be away for a known period of time.
>
> Please feel free to ping me (@srmnitc ) if you have any questions/concerns.
