### 2024-02-16 #6358 @Extraweich
@julian-belina there is a „of“ too much in the first sentence in your paper. Would there be a possibility to make a reference to the ETHOS project?
