### 2024-02-16 #6358 @julian-belina
@Extraweich Thank you for pointing out the mistake in the first sentence. I will fix this later.  I am not sure if there is a persistent link to ETHOS but I will figure this out. In the meantime there is this link:
https://www.fz-juelich.de/de/iek/iek-3/leistungen/model-services
