### 2024-02-17 #6358 @Extraweich
@julian-belina concerning the installation guide, your "Mamba installation guide" does not work for me. Could you check this please?
