### 2024-02-19 #6358 @julian-belina
@Extraweich The surrplus word has been removed from the paper, a hyperlink to ETHOS has been added to the paper and the link to the mamba installation has been fixed.

Thank you for your feedback so far. I am looking forward for any remarks you might have.
