### 2024-02-19 #6358 @Extraweich
@julian-belina Thank you for fixing the issues. Currently I am still stuck in the installation. If I follow the description on your repository, I get the following error:

Could you please check this and adjust your installation guide?
