### 2024-02-19 #6358 @julian-belina
@Extraweich Could you send me the full error message?
Also the the ouputs of the conda info would be helpfull. It seems to be related to your operating system.
