### 2024-02-19 #6358 @Extraweich
This was the full error message. Concerning `conda info`, this is the output

If I try to install pdftopng using `pip install pdftopng`, I get the following:

```
(penalps_env)` ...~ % ./miniforge3/bin/pip3 install pdftopng
Collecting pdftopng
  Using cached pdftopng-0.1.0.tar.gz (11 kB)
  Installing build dependencies ... done
  Preparing metadata (pyproject.toml) ... done
Collecting Click>=7.0 (from pdftopng)
  Using cached click-8.1.7-py3-none-any.whl.metadata (3.0 kB)
Using cached click-8.1.7-py3-none-any.whl (97 kB)
Building wheels for collected packages: pdftopng
  Building wheel for pdftopng (pyproject.toml) ... error
  error: subprocess-exited-with-error

  × Building wheel for pdftopng (pyproject.toml) did not run successfully.
  │ exit code: 1
  ╰─> [35 lines of output]
      running bdist_wheel
      running build
      running build_py
      creating build
      creating build/lib.macosx-11.0-arm64-cpython-310
      creating build/lib.macosx-11.0-arm64-cpython-310/pdftopng
      copying src/pdftopng/__init__.py -> build/lib.macosx-11.0-arm64-cpython-310/pdftopng
      copying src/pdftopng/__version__.py -> build/lib.macosx-11.0-arm64-cpython-310/pdftopng
      running egg_info
      writing src/pdftopng.egg-info/PKG-INFO
      writing dependency_links to src/pdftopng.egg-info/dependency_links.txt
      writing requirements to src/pdftopng.egg-info/requires.txt
      writing top-level names to src/pdftopng.egg-info/top_level.txt
      reading manifest file 'src/pdftopng.egg-info/SOURCES.txt'
      writing manifest file 'src/pdftopng.egg-info/SOURCES.txt'
      copying src/pdftopng/pdftopng.cpp -> build/lib.macosx-11.0-arm64-cpython-310/pdftopng
      running build_ext
      creating var
      creating var/folders
      creating var/folders/fp
      creating var/folders/fp/lk98qxrs3kb4pcphb21ssdjr0000gn
      creating var/folders/fp/lk98qxrs3kb4pcphb21ssdjr0000gn/T
      clang -Wno-unused-result -Wsign-compare -Wunreachable-code -DNDEBUG -fwrapv -O2 -Wall -fPIC -O2 -isystem /Users/nicolaschrist/miniforge3/include -arch arm64 -fPIC -O2 -isystem /Users/nicolaschrist/miniforge3/include -arch arm64 -I/Users/nicolaschrist/miniforge3/include/python3.10 -c /var/folders/fp/lk98qxrs3kb4pcphb21ssdjr0000gn/T/tmpadyagopp.cpp -o var/folders/fp/lk98qxrs3kb4pcphb21ssdjr0000gn/T/tmpadyagopp.o -std=c++14
      clang -Wno-unused-result -Wsign-compare -Wunreachable-code -DNDEBUG -fwrapv -O2 -Wall -fPIC -O2 -isystem /Users/nicolaschrist/miniforge3/include -arch arm64 -fPIC -O2 -isystem /Users/nicolaschrist/miniforge3/include -arch arm64 -I/Users/nicolaschrist/miniforge3/include/python3.10 -c /var/folders/fp/lk98qxrs3kb4pcphb21ssdjr0000gn/T/tmpczjqy2yw.cpp -o var/folders/fp/lk98qxrs3kb4pcphb21ssdjr0000gn/T/tmpczjqy2yw.o -fvisibility=hidden
      building 'pdftopng.pdftopng' extension
      creating build/temp.macosx-11.0-arm64-cpython-310
      creating build/temp.macosx-11.0-arm64-cpython-310/src
      creating build/temp.macosx-11.0-arm64-cpython-310/src/pdftopng
      clang -Wno-unused-result -Wsign-compare -Wunreachable-code -DNDEBUG -fwrapv -O2 -Wall -fPIC -O2 -isystem /Users/nicolaschrist/miniforge3/include -arch arm64 -fPIC -O2 -isystem /Users/nicolaschrist/miniforge3/include -arch arm64 -DVERSION_INFO=\"0.1.0\" -Ilib/poppler -Ilib/poppler/fofi -Ilib/poppler/goo -Ilib/poppler/poppler -Ilib/poppler/build -Ilib/poppler/build/poppler -Ilib/poppler/utils -Ilib/poppler/build/utils -I/private/var/folders/fp/lk98qxrs3kb4pcphb21ssdjr0000gn/T/pip-build-env-potoll2p/overlay/lib/python3.10/site-packages/pybind11/include -I/Users/nicolaschrist/miniforge3/include/python3.10 -c src/pdftopng/pdftopng.cpp -o build/temp.macosx-11.0-arm64-cpython-310/src/pdftopng/pdftopng.o -O3 -Wall -shared -fPIC -stdlib=libc++ -mmacosx-version-min=10.7 -std=c++14 -fvisibility=hidden
      clang: warning: argument unused during compilation: '-shared' [-Wunused-command-line-argument]
      src/pdftopng/pdftopng.cpp:44:10: fatal error: 'config.h' file not found
      #include "config.h"
               ^~~~~~~~~~
      1 error generated.
      error: command '/usr/bin/clang' failed with exit code 1
      [end of output]

  ERROR: Failed building wheel for pdftopng
Failed to build pdftopng
ERROR: Could not build wheels for pdftopng, which is required to install pyproject.toml-based projects
```

I don't think this is an issue with Apple silicon. I removed the mamba installation and tried installing pdftopng, which then worked in my conda environment.
