### 2024-02-19 #6358 @julian-belina
I did not mean that your operating system if faulty ot that pdftopng is not isntallable on your system. Conda-forge is a Python Software repository that is diffrent from pypi which is used by pip. It is possible that the for you combination of python version and operating system there is no version of pdftopng available. The error message of conda can be quite misleading.

Generally it should be available on conda forge:
https://anaconda.org/conda-forge/pdftopng

If the error persists i will try to use a different dependency. Pdftopng is not that important for my library.
