### 2024-02-19 0874b1a
remove large macos runner
