+++ .github/workflows/test.yml
@@ -22,7 +22,7 @@ jobs:
       strategy:
         fail-fast: false
         matrix:
-          os: ["ubuntu-latest","ubuntu-20.04", "macos-latest","macos-13","macos-13-xlarge","macos-12", "macos-11","windows-latest","windows-2019"]
+          os: ["ubuntu-latest","ubuntu-20.04", "macos-latest","macos-13","macos-12", "macos-11","windows-latest","windows-2019"]
       steps:
         - name: Checkout
           uses: actions/checkout@v4
@@ -53,7 +53,7 @@ jobs:
       strategy:
         fail-fast: false
         matrix:
-          os: ["ubuntu-latest","ubuntu-20.04", "macos-latest","macos-13","macos-13-xlarge","macos-12", "macos-11","windows-latest","windows-2019"]
+          os: ["ubuntu-latest","ubuntu-20.04", "macos-latest","macos-13","macos-12", "macos-11","windows-latest","windows-2019"]
       steps:
         - name: Checkout
           uses: actions/checkout@v4
