### 2024-02-19 #6358 @julian-belina
@srmnitc @Extraweich I have created an issue at the respective conda-forge feedstock, which provides the respective conda-forge package:
https://github.com/conda-forge/pdftopng-feedstock/issues/20

The installation fails because conda-forge only provides the package for osx-64 and not for osx-arm64.

Does Joss require installability on each operating system and processor combination?
