### 2024-02-21 #6358 @srmnitc
> The installation fails because conda-forge only provides the package for osx-64 and not for osx-arm64.
>
> Does Joss require installability on each operating system and processor combinations?

@julian-belina this is not necessarily needed. But the package pdftopng has not been maintained for 3 years now. So maybe it might be better for your package to explore other options.
