### 2024-02-21 5e39a78
Adjust Github CI and remove pdftopng
