### 2024-02-21 #6358 @julian-belina
@srmnitc  @Extraweich pdftopng has been removed in version 1.0.2 of ethos_penalps. The conda-forgen package should be installable within the next two hours.
