### 2024-02-25 FZJ-IEK3-VSA/ETHOS_PeNALPS#2 @stoldtjo
Link do docs broken in readme.md
The link on the readme.md starting page of the project is broken and clicking it will cause the user to be referred to a 404 error page.
