### 2024-02-25 FZJ-IEK3-VSA/ETHOS_PeNALPS#2 @stoldtjo
Links in section [Working Principle](https://github.com/FZJ-IEK3-VSA/ETHOS_PeNALPS/tree/main?tab=readme-ov-file#working-principle) are also broken.
