### 2024-02-25 #6358 @stoldtjo
@julian-belina: Couple of remarks and comments at this point:

1. There is an issue with links to the docs in the readme.md: https://github.com/FZJ-IEK3-VSA/ETHOS_PeNALPS/issues/2
2. With respect to authorship: You are the only contributor to the repository of ETHOD.PeNALPS. Yet, authorship is shared with two more individuals without the paper specifying their respective contributions. Some explanation should be shared.
3. Documentation: The docs at https://ethospenalps.readthedocs.io/en/latest/contents.html are providing little in addition to what is in the paper, except for the presumably auto-generated API docs. SimPy, for instance, has a section "[SimPy in 10 Minutes](https://simpy.readthedocs.io/en/latest/simpy_intro/index.html)", which seems like an easier starting point compared to the more complex examples provided by ETHOS.PeNALPS.
4. General thought on the implementation: Many previous works built on existing simulation engines/software to leverage their existing distribution and level of development. Why has ETHOS.PeNALPS chosen to start from scratch rather than try and extend SimPy or some some other FOSS simulation package?
5. API Docs: I usually work a lot with Microsoft docs and I used to work with Plant Simulation references. Compared to either the API Docs are lacking. Without checking the example, doing some elaborate guess work or actually looking into the project's code I would feel somewhat lost which methods and attributes are responsible for which purpose and what I should expect from them in terms of functionality. ethos_penalps.process_nodes.process_step, for instance, is a little better than the majority of samples I checked (e.g., ethos_penalps.simulation_data.container_simulation_data).
6. asd.py: What is that file supposed to do? Judging from the file name it feels like some personal test that wasn't cleaned from the repository.
7. Community guidelines for contributions from other developers and users are not provided (cf. checklist).
8. Paper structure: The order of sections is rather unconventional. Particularly the fact that the discussoin of other works is at the end of the paper makes that section feel a tad random. Generally, the structure should follow a logic stream that usually goes along the lines of "Motivation and problem definition, review of state of the art, concept, implementation, validation on a use case, conclusion".
9. Paper Section "Statement of Need" is weak because what has been noted as a requirement is kept very generic and general. Particularly compared with the implemented energy model the software is capable of serving the need. The implemented energy consumption model (average consumption per object state) is suitable for a small set of potential analysis questions but many more would hardly profit from this modeling approach. See also this chapter: http://dx.doi.org/10.1007/978-3-031-34218-9_1
10. Paper Section "Other Tools and Methods" is weak and fails to give a solid overview of other tools and methods. A number of review papers have been published and they provide some additional pointers on relevant works and publications. Particularly early works in the field from between 2005 and 2015 contained a number of works that report on non-commercial tools developed by various research group. Authors Wohlgemuth, Pawletta, Andersson, Grundmeier, Heilala, Rippel, Sobottka come to mind. It might suffice to reference appropriate review papers but it definitely seems advisable that previous works are checked for their licensing status to ensure the statement "most of these are not open source" is not just driven by those works which are based on commercial tools.
11. Paper Discussion of Modeling approach: The fact that energy demand is modelled as an average consumption per state should be discussed in the paper. There are a number of implications to the use of the software originating from that modelling approach which are worthwhile to elaborate.
12. Paper List of References: The selection of references feels rather arbitrary to someone who has been following this field of research for some years. The authors should either attempt to present a more complete picture of existing research/publications or refer to publications providing it and commenting why specific examples were selected and how they compare to other non-cited publications (e.g., to elaborate a specific example in a little more depth while mentioning that the example stands for a group of similar approaches).
13. Paper Summary: The paper misses any conclusion or summary.

All considered, I believe the documentation of the software (API and modelling approach) as well as the paper should undergo major revision before the submission is accepted for publication.
