### 2024-02-25 FZJ-IEK3-VSA/ETHOS_PeNALPS#3 @stoldtjo
Energy plots look strange (Date missing?)
I ran the toffee production example and found the generated energy charts to be strange. There are two categories on the x axis, on 2022-12-31 and one without description. It looks like some datapoints going into that image are attributed to a blank date. Running on Windows 11 x64 on fresh Miniforge3 installation.
![process_state_load_profile_Toffee Machine 1-Electricity](https://github.com/FZJ-IEK3-VSA/ETHOS_PeNALPS/assets/118894305/0a679c0e-bc90-4071-9b3c-cf80efdecf39)
![process_state_load_profile_Toffee Machine 1-Natural Gas](https://github.com/FZJ-IEK3-VSA/ETHOS_PeNALPS/assets/118894305/2413e5a5-3f8e-48de-8727-765527abd763)
![process_state_load_profile_Toffee Machine 2-Electricity](https://github.com/FZJ-IEK3-VSA/ETHOS_PeNALPS/assets/118894305/8bde4fe4-45c2-414b-8e7d-d1bb205486e9)
![process_state_load_profile_Toffee Machine 2-Natural Gas](https://github.com/FZJ-IEK3-VSA/ETHOS_PeNALPS/assets/118894305/59f56937-31ac-4e9d-ba20-a27130ef30be)
![stream_load_profile_Toffee_Raw_Materials_Toffee_Machine_2_Raw_Toffee_Ingredients-Electricity](https://github.com/FZJ-IEK3-VSA/ETHOS_PeNALPS/assets/118894305/ecd2386d-5c11-4807-adac-d373e8e44c38)
![summary_load_profile-Electricity](https://github.com/FZJ-IEK3-VSA/ETHOS_PeNALPS/assets/118894305/2d592261-b85f-45d0-8adf-8f39940e6b15)
![summary_load_profile-Natural Gas](https://github.com/FZJ-IEK3-VSA/ETHOS_PeNALPS/assets/118894305/775e644d-9b38-4473-b9e4-f66ef977458d)
![total_energy_load_profile](https://github.com/FZJ-IEK3-VSA/ETHOS_PeNALPS/assets/118894305/7aa507cf-0de2-40f4-83cc-5d2e4594227a)
