### 2024-02-25 FZJ-IEK3-VSA/ETHOS_PeNALPS#3 @stoldtjo
Incidentally, they do look different on the B-Pillar example... No mention of 2022-12-31 there.
