### 2024-02-26 eca85b5
Add authors contribution, fix links in readme
