### 2024-02-27 343e80d
Add tutorail and workspace recommendations
