### 2024-02-27 #6358 @stoldtjo
Dear @julian-belina,

Thank you for your response. Below some comments to your remarks.

> Concerning your points 11 - 14, I wanted to mention that [the target JOSS paper length is between 250 and 1000 words](https://joss.readthedocs.io/en/latest/submitting.html#what-should-my-paper-contain). Because I already exceeded this limit by 700 words, I am in a conflict of completenes and text length.
Thanks for pointing that out - I wasn't aware. I understand that such a sharp limitation is difficult. Nevertheless, I believe the case study/example can be cut short or mostly omitted (with reference to the docs of the project). That should free some space to properly restructure the paper and build a consistent stream in the presentation.

> Concerning point 11 and 13, I agree that these publications form an important basis for the subject matter. But I think that any attempt to give a rather complete picture of the previous publications will go beyond the target text length of a Joss publication. Unfortunately, I am not aware of a suitable review that focuses on the simulation of industrial manfufacturing processes with temporal resolution besides the one from [Grimm et al.](https://doi.org/10.1007/978-3-030-90700-6_62). This review is already outdated and does not consider [Binderbauer et al.](https://doi.org/10.1016/j.jclepro.2021.130024), [Dock et al.](https://doi.org/10.1007/s00502-021-00895-0) or [Sandhaas et al.](https://doi.org/10.3390/en15103683). Also the search string of the systematic review was incomplete. This lead to the false conclusion that there are no publications that simulate load profiles of industrial manufacturing processes using bottom-up methods.
>
> My current selection focusses on publications which formulate the objective to simulate and publish load profiles of industrial manufacturing procces. This data is of particular interest in the field of energy system design and analysis to quantify the need for energy flexibility measures. Previous method publications such as the one from [Kohl et al.](https://doi.org/10.1016/j.procir.2014.01.088) simulate load profiles as an intermediate result, which I only discovered by conducting a backwards search from [Garwood et al.](https://doi.org/10.1016/j.rser.2017.08.063). However, the review by Garwood et al. has a much larger scope which made me question the suitability for this paper.
>
> The review by [Walther et al.](https://doi.org/10.3390/en14040968) only deals with the prediction and forecasting of electrical energy consumption of the manufacturing industry. It does not consider other energy carriers and the temporal resolution of the predicted data. Also it does not consider the simulation of load profiles of processes in the past.
>
> Therefore, I would like to seek your opinion. Do you believe that an extension of the scope of the literature is necessary, or do you see ways in which the brevity of the text can be maintained without losing quality or informational content? I am open to any suggestions and would appreciate hearing your specific recommendations.
Wenzel et al. (2018) presented at ASIM SPL 2017 conference mentions a number of review papers that could be referenced directly, all of which are focussed on highly relevant works. The ASIM SPL workgroup BeESPL has furthermore published at WSC 2018 and ASIM SPL 2021. My own dissertation ([available in Open Access](https://nbn-resolving.org/urn:nbn:de:bsz:ch1-qucosa2-343579)) also has a rather extensive review section on the combined simulation of material and energy flows in industrial processes.

Anyhow, to me this is not so much about reproducing or summarizing all the many previous works or having a review that was published very recently. I find it more suitable to reference relevant reviews - after careful study, of course - and then maybe select some exceptionally works with some argument for their outstanding relevance beyond other works for your project. An alternative approach would be to follow some kind of classification of the properties of published studies but that's actually what the ASIM SPL workgroup BeESPL did in their publications. Eventually it becomes a question of "Why select those specific references and how do they contribute to the motivation for the work on ETHOS.PeNALPS?" and the paper
s text should provide the answer to that.

> To 12, modeling rising, falling and fluctuating energy demands for individual states is planned for a future release of this software. Currently it is not implemented due to time constraints. However, it is already possible to add an arbitrary number of states with different specific energy demands. This provides the flexibility to model complex energy demand patterns. Still, this should be discussed in the paper.
Actually, both the planned extension and also the implications of the currently available model should be part of the paper. In terms of implications I suggest you focus specifically on "What does that mean in terms of potential use cases?" As you are aware, the lack in resolution that is inherent to steady consumption per state disqualify the solution for most very technical considerations but are quite sufficient for economic evaluations, e.g., concerning the costs for electricity based on spot market prices. Relevant limitations should be made clear in the presentation of the product.

The above exchange was mostly focussed on the paper so far. Especially considering the length constraints and the overall JOSS review guidelines I want to highlight the importance of also working on the suggestions to the repository and especially the docs I have provided previously.

Kind regards,
Johannes Stoldt
