### 2024-02-28 #6358 @srmnitc
@julian-belina In case the length of paper is not sufficient to fully express the background, you can also make use of your documentation pages, and include a shorter version in the paper.
