+++ documentation/ethos_penalps_tutorial/add_network_level_and_process_chains.md
@@ -1,6 +1,0 @@
-# Add more network level and process chains
-
-This sections explains how multiple process steps can be used in the simulation. They can be organized either sequentially or parallelly. Process steps that should be organized sequentially. 
-
-
-Multiple network level are used to model sequential process steps. Multiple process chains within a network level are used to model parallel operation of process steps. In the case of a single machine an instance of each is sufficient.  
