+++ examples/tutorial/_2_cooking_example_more_states.py
@@ -1,0 +1,235 @@
+import datetime
+import logging
+
+from ethos_penalps.data_classes import Commodity, LoadType
+from ethos_penalps.enterprise import Enterprise
+from ethos_penalps.order_generator import NOrderGenerator
+from ethos_penalps.stream import BatchStreamStaticData, ContinuousStreamStaticData
+from ethos_penalps.time_data import TimeData
+from ethos_penalps.utilities.logger_ethos_penalps import PeNALPSLogger
+
+logger = PeNALPSLogger.get_human_readable_logger(logging.INFO)
+
+# Enterprise structure
+
+# Set simulation time data
+start_date = datetime.datetime(2022, 1, 2, hour=22, minute=30)
+end_date = datetime.datetime(2022, 1, 3)
+time_data = TimeData(
+    global_start_date=start_date,
+    global_end_date=end_date,
+)
+
+
+# Determine all relevant commodities
+output_commodity = Commodity(name="Cooked Goods")
+input_commodity = Commodity(name="Raw Goods")
+
+
+# Create all order for the simulation
+order_generator = NOrderGenerator(
+    commodity=output_commodity,
+    mass_per_order=0.0005,
+    production_deadline=end_date,
+    number_of_orders=2,
+)
+
+order_collection = order_generator.create_n_order_collection()
+
+# Initialize enterprise
+enterprise = Enterprise(time_data=time_data, name="Cooking Example")
+
+# Create network level
+network_level = enterprise.create_network_level()
+# Create first process chain
+
+process_chain = network_level.create_process_chain(process_chain_name="Cooker Chain")
+
+# Create all sources, sinks and network level storages
+sink = network_level.create_main_sink(
+    name="Cooked Goods Storage",
+    commodity=output_commodity,
+    order_collection=order_collection,
+)
+source = network_level.create_main_source(
+    name="Raw Material Storage",
+    commodity=input_commodity,
+)
+
+
+# Add sources and sinks to process chain
+process_chain.add_sink(sink=sink)
+process_chain.add_source(source=source)
+
+# Create Process nodes
+process_step = process_chain.create_process_step(name="Cooker")
+
+# Streams
+raw_materials_to_cooking_stream = process_chain.stream_handler.create_batch_stream(
+    batch_stream_static_data=BatchStreamStaticData(
+        start_process_step_name=source.name,
+        end_process_step_name=process_step.name,
+        delay=datetime.timedelta(minutes=1),
+        commodity=input_commodity,
+        maximum_batch_mass_value=0.0005,
+    )
+)
+cooking_to_sink_stream = process_chain.stream_handler.create_batch_stream(
+    batch_stream_static_data=BatchStreamStaticData(
+        start_process_step_name=process_step.name,
+        end_process_step_name=sink.name,
+        delay=datetime.timedelta(minutes=1),
+        commodity=output_commodity,
+        maximum_batch_mass_value=0.0005,
+    )
+)
+
+# Add streams to sinks and sources
+source.add_output_stream(
+    output_stream=raw_materials_to_cooking_stream,
+    process_chain_identifier=process_chain.process_chain_identifier,
+)
+sink.add_input_stream(
+    input_stream=cooking_to_sink_stream,
+    process_chain_identifier=process_chain.process_chain_identifier,
+)
+
+""" Create petri net for process step
+Each process state must have at least the following:
+- Either
+    - one combined production state
+    your_combined_state=process_step.process_state_handler.create_continuous_production_process_state_with_storage(
+    process_state_name="your process state name"
+    )
+    or
+        - an input stream requesting state
+        your_input_stream_requesting_state=process_step.process_state_handler.create_continuous_input_stream_requesting_state(process_state_name="your input stream providing state")
+        - and input stream requesting state
[…]
