+++ src/ethos_penalps/utilities/units.py
@@ -5,10 +5,10 @@ import pint
 class Units:
     unit_registry = pint.UnitRegistry(system="SI")
     unit_registry.default_format = "~"
-    time_unit: pint.Unit = unit_registry.seconds
+    time_unit: pint.Unit = unit_registry.hour
     mass_unit: pint.Unit = unit_registry.metric_ton
-    power_unit: pint.Unit = unit_registry.watt
-    energy_unit: pint.Unit = unit_registry.joule
+    power_unit: pint.Unit = unit_registry.MW
+    energy_unit: pint.Unit = unit_registry.MJ
 
     def get_unit(unit_string: str) -> pint.Unit:
         return Units.unit_registry.Unit(unit_string)
@@ -22,3 +22,12 @@ class Units:
 
     def get_value_from_quantity(quantity: pint.Quantity) -> numbers.Number:
         return quantity.m
+
+
+if __name__ == "__main__":
+    unit_registry = pint.UnitRegistry(system="SI")
+
+    meter_quant = (0.15 * Units.unit_registry.Unit("kWh")) / (
+        650 * Units.unit_registry.Unit("gram")
+    )
+    print(meter_quant.to("MJ/metric_ton"))
