+++ pyproject.toml
@@ -2,6 +2,10 @@
 requires = ["setuptools>=64.0.0", "wheel"]
 build-backend = "setuptools.build_meta"
 
+[tool.setuptools.package-data]
+"ethos_penalps" = ["py.typed"]
+[tool.setuptools.packages.find]
+where = ["src"]
 
 [project]
 name = "ethos_penalps"

[…]
