### 2024-02-29 428ae39
Recompile paper, add community guidelines for contributions
