### 2024-02-29 #6358 @julian-belina
@stoldtjo

  > Wenzel et al. (2018) presented at ASIM SPL 2017 conference mentions a number of review papers that could be referenced directly, all of which are focussed on highly relevant works. The ASIM SPL workgroup BeESPL has furthermore published at WSC 2018 and ASIM SPL 2021. My own dissertation ([available in Open Access](https://nbn-resolving.org/urn:nbn:de:bsz:ch1-qucosa2-343579)) also has a rather extensive review section on the combined simulation of material and energy flows in industrial processes.

>   Anyhow, to me this is not so much about reproducing or summarizing all the many previous works or having a review that was published very recently. I find it more suitable to reference relevant reviews - after careful study, of course - and then maybe select some exceptionally works with some argument for their outstanding relevance beyond other works for your project. An alternative approach would be to follow some kind of classification of the properties of published studies but that's actually what the ASIM SPL workgroup BeESPL did in their publications. Eventually it becomes a question of "Why select those specific references and how do they contribute to the motivation for the work on ETHOS.PeNALPS?" and the papers text should provide the answer to that.

Thank you for you literature recommendations. I will read them in the next days.

Concerning your previous points:

- 1 and 6 have been fixed.
- 2.  An authors contribution section has been added to the paper.
- 3.  A tutorial has been added to repository and documentation.
- 4. I have decided not to use simpy for the following reasons:
   - The discrete event simulation is of ETHOS.PeNALPS is conducted backwards from a start point to the past. This is not implemented in simpy.
   - Simpy uses a uniform time grid, whereas ETHOS.PeNALPS is based on a non uniform time grid. This prevents decisions to be dependent on the temporal resolution of the time grid
   - The discrete event simulation is only a subset of the ETHOS.PeNALPS simulation. In the beginning it was not clear if it would cause problem to embbed the simpy as it is.
- 8. Community Guidelines have been added to the Readme.

Concerning point 5 and 7. I will continue to improve the API description and work on the reported bug.

@srmnitc  @stoldtjo
Concerning point 9-14, I dont think that it is usefull to do a full literature review in the HTML documentation. I have to read into reviews before I can make a proposal to improve the literature part and structure of the paper.
