### 2024-02-29 #6358 @stoldtjo
Hello @julian-belina,

I skimmed through the tutorial. Well done, exactly what I would have expected. Unfortunately, I don't have the time to try and make use of it on a new example but I trust you there are no issues in the code provided.

Regarding you decision against SimPy: Your first point seems of little relevance to me, to be honest. Since your project does not provide an interactive mode or a playback mode it's of no consequence to the user. I don't quite understand the comment about time grid. Discrete event simulation to me means that simulation time follows events (state changes) and not vice versa. Anyhow, I am not an expert on SimPy so I may be missing something in terms of implementation. Anyhow, no further questions on that point. It was mostly curiosity to boot with.

Community Guidelines looks nice, too.

Kind regards,
Johannes Stoldt
