### 2024-03-01 FZJ-IEK3-VSA/ETHOS_PeNALPS#2 @julian-belina
Thank you for pointing out the bug. It has been fixed.
