+++ README.md
@@ -18,8 +18,7 @@ objects. After the material flow simulation is completed, a set of production or
 ![Main Component Overview](paper/main_component_overview.png)
 *Depiction of the main components and workflow of ETHOS.PeNALPS*
 
-A further description of the model definition can be found [here]([ethos_penalps_articles/model_description.md](https://ethospenalps.readthedocs.io/en/latest/ethos_penalps_articles/model_description.html)). 
-Also two examples for a [toffee production process](https://ethospenalps.readthedocs.io/en/latest/examples/toffee_example.html) and a [b-pillar production process](https://ethospenalps.readthedocs.io/en/latest/examples/b_pillar_example.html) are available.
+The [HTML documentation provides a tutorial](https://ethospenalps.readthedocs.io/en/latest/ethos_penalps_tutorial/overview.html) for ETHOS.PeNALPS. The executable files for the tutorial are located in the example section of this repository. Also two examples for a [toffee production process](https://ethospenalps.readthedocs.io/en/latest/examples/toffee_example.html) and a [b-pillar production process](https://ethospenalps.readthedocs.io/en/latest/examples/b_pillar_example.html) are available.
 
 
 # Installation
@@ -29,7 +28,7 @@ The installation process uses a Conda-based Python package manager. We highly re
 
 
 ## Installation via conda-forge
-The simplest way ist to install FINE into a fresh environment from conda-forge with:
+The simplest way ist to install ETHOS.PeNALPS into a fresh environment from conda-forge with:
 
 Create a new environment
 ```python
