### 2024-03-02 FZJ-IEK3-VSA/ETHOS_PeNALPS#2 @stoldtjo
Hi, there are still errors. See links in following sentences, both of which yield 404.

> A further description of the model definition can be found [here](https://github.com/FZJ-IEK3-VSA/ETHOS_PeNALPS/blob/main/%5Bethos_penalps_articles/model_description.md%5D(https:/ethospenalps.readthedocs.io/en/latest/ethos_penalps_articles/model_description.html)).

> A ReadTheDocs Documentation can be found [here](https://ethospenalps.readthedocs.io/en/latest/Introduction.html).
