### 2024-03-03 #6358 @dpzhuX
@julian-belina I've successfully installed the package and run several examples provided in the package. I have a few questions and observations related to the package functionality and documentation:

1. Validation of examples: There are two examples mentioned in the package documentation. Is there any validation available for these examples? Additionally, I noticed that the results for these examples are not displayed on the documentation website.
2. Load profiles: How were the load profiles (c) generated in Figure 2? After running the `toffee_production` example, I saw the simulation report but did not find load profiles similar to those depicted in load profiles (c) of Figure 2.
3. Parallel chain design: For the parallel chain design in `toffee_production` example, I observed that the primary distinction between chain 1 and chain 2 seems to be their names (i.e., Machine 1 vs Machine 2). If we want to have numerous parallel chains (e.g., 1000 chains) with identical designs, is it necessary to manually duplicate and rename each file for every chain? If this is the case, I think the code design might need to improve.
4. Multiple sources and sinks: A general question for this package that I did not find answer in the documentation. The documentation does not explicitly address whether a chain can include multiple sources and sinks for a complex topology. Can you confirm if the package can accommodate such a configuration:
```
   +----Source----+
   |              |
   v              v
Machine2 <-- Machine1
   |              |
   |              |
   +-> Storage <--+
```
5. Issue with running an example: I have an issue when attempting to run the  `_4_connect_four_process_steps` from the tutorial folder, I have the following error:

```
(ethos_penalps) dzhu:~/Downloads/ETHOS_PeNALPS/examples/tutorial/_4_connect_four_process_steps$python simulation_starter.py
Traceback (most recent call last):
  File "/home/dzhu/Downloads/ETHOS_PeNALPS/examples/tutorial/_4_connect_four_process_steps/simulation_starter.py", line 10, in <module>
    from examples.tutorial._4_connect_four_process_steps.blending_process_chain import (
ModuleNotFoundError: No module named 'examples'
```
