+++ test/test_carpet_plot/asd.py
@@ -1,0 +1,7 @@
+import itertools
+
+a = itertools.count()
+b = itertools.cycle([3, 4, 5])
+
+for i in range(6):
+    print(next(b))
