+++ src/ethos_penalps/post_processing/load_profile_post_processor.py
@@ -331,53 +331,44 @@ class LoadProfilePostProcessor:
         figure, axes = matplotlib.pyplot.subplots(1, 1)
         axes.minorticks_off()
         converted_index = []
-        load_profile_matrix = carpet_plot_load_profile_matrix.data_frame
+        load_profile_matrix_data_frame = carpet_plot_load_profile_matrix.data_frame
         start_date = carpet_plot_load_profile_matrix.start_date_time_series
         end_date = carpet_plot_load_profile_matrix.end_date_time_series
         x_axis_whole_period = (
             carpet_plot_load_profile_matrix.x_axis_time_period_timedelta
         )
         resample_frequency = carpet_plot_load_profile_matrix.resample_frequency
-        for index_entry in load_profile_matrix.index:
+        for index_entry in load_profile_matrix_data_frame.index:
             if isinstance(index_entry, datetime.datetime):
                 a = index_entry.to_pydatetime()
-                # print(a, type(a))
                 py_date_time_entry = index_entry.to_pydatetime(a)
-                # cut_py_date_time_entry = py_date_time_entry.time()
-                # converted_index.append(py_date_time_entry.strftime(y_axis_time_format))
-                converted_index.append(py_date_time_entry)
+                converted_index.append(str(py_date_time_entry))
             else:
                 converted_index.append(index_entry)
-        load_profile_matrix.index = converted_index
+        load_profile_matrix_data_frame.index = converted_index
+        converted_row_index = []
+        for row_entry in load_profile_matrix_data_frame.columns:
+            if isinstance(row_entry, datetime.datetime):
+                a = index_entry.to_pydatetime()
+                py_date_time_entry = row_entry.to_pydatetime(a)
+                converted_row_index.append(
+                    str(py_date_time_entry).strftime(y_axis_time_format)
+                )
+            else:
+                converted_row_index.append(index_entry)
+        load_profile_matrix_data_frame.columns = converted_row_index
 
-        load_profile_matrix_numpy = load_profile_matrix.to_numpy()
+        load_profile_matrix_numpy = load_profile_matrix_data_frame.to_numpy()
 
         searborn_plot = seaborn.heatmap(
-            load_profile_matrix_numpy,
+            load_profile_matrix_data_frame,
             cmap="coolwarm",
             ax=axes,
-            # vmin=0,
             center=0,
-            # yticklabels=converted_index,
-            # yticklabels=False,
             cbar_kws={"label": "Average Power in MW"},
+            xticklabels="auto",
         )
-        timedelta_frequency = pandas.to_timedelta(resample_frequency)
-        total_energy_demand = self.get_energy_amount_from_energy_plotting_matrix(
-            load_profile_data_frame=load_profile_matrix,
-        )
-        total_energy_quantity = Units.compress_quantity(
-            unit=Units.energy_unit, quantity_value=total_energy_demand
-        )
-
-        matplotlib.pyplot.title(
-            label=str(round(total_energy_quantity.m, 2))
-            + " "
-            + create_subscript_string_matplotlib(
-                base=str(total_energy_quantity.u),
-                subscripted_text=str(load_type_name),
-            )
-        )
+        # timedelta_frequency = pandas.to_timedelta(resample_frequency)
 
         number_of_y_rows = len(converted_index)
         if x_axis_whole_period <= datetime.timedelta(hours=1):
@@ -408,16 +399,16 @@ class LoadProfilePostProcessor:
             lin_loactor = matplotlib.ticker.LinearLocator(
                 22,
[…]
