+++ src/ethos_penalps/post_processing/load_profile_post_processor.py
@@ -1,6 +1,4 @@
 import datetime
-import math
-
 
 import matplotlib.dates
 import matplotlib.figure
@@ -10,35 +8,29 @@ import numpy
 import pandas
 import seaborn
 
-from dataclasses import dataclass
 from ethos_penalps.data_classes import (
+    CarpetPlotMatrix,
+    CarpetPlotMatrixEmpty,
     LoadProfileDataFrameMetaInformation,
     LoadProfileEntry,
     LoadType,
+    ListOfLoadProfileMetaData,
+    ListLoadProfileMetaDataEmpty,
 )
-from ethos_penalps.utilities.units import Units
 from ethos_penalps.post_processing.load_profile_entry_post_processor import (
     LoadProfileEntryPostProcessor,
+    ListOfLoadProfileEntryAnalyzer,
 )
-
-from ethos_penalps.utilities.units import Units
+from ethos_penalps.utilities.exceptions_and_warnings import Misconfiguration
 from ethos_penalps.utilities.general_functions import (
     check_if_date_1_is_before_date_2,
     check_if_date_1_is_before_or_at_date_2,
     create_subscript_string_matplotlib,
 )
 from ethos_penalps.utilities.logger_ethos_penalps import PeNALPSLogger
+from ethos_penalps.utilities.units import Units
 
-itcs_logger = PeNALPSLogger.get_logger_without_handler()
-
-
-@dataclass
-class CarpetPlotMatrix:
-    data_frame: pandas.DataFrame
-    start_date_time_series: datetime.datetime
-    end_date_time_series: datetime.datetime
-    x_axis_time_period_timedelta: datetime.timedelta
-    resample_frequency: str = "1min"
+logger = PeNALPSLogger.get_logger_without_handler()
 
 
 class LoadProfilePostProcessor:
@@ -57,20 +49,54 @@ class LoadProfilePostProcessor:
         list_of_load_profile_entries: list[LoadProfileEntry],
         start_date_time_series: datetime.datetime,
         end_date_time_series: datetime.datetime,
+        object_name: str,
         x_axis_time_period_timedelta: datetime.timedelta = datetime.timedelta(weeks=1),
         resample_frequency: str = "1min",
-    ) -> CarpetPlotMatrix:
+    ) -> CarpetPlotMatrix | CarpetPlotMatrixEmpty:
+        """Converts a list of load profiles into a CarpetPlotMatrix
+        which is used to create carpet plots.
+            1. The list of load profiles is homogenized
+            2. The list of load profiles is converted into a data frame
+            3. The data frame and meta information about the list
+                is passed to CarpetPlotMatrix instance.
+
+        Args:
+            list_of_load_profile_entries (list[LoadProfileEntry]): The list of load profiles
+            start_date_time_series (datetime.datetime): The new start time of the list of load profiles.
+                If the start time is later than the earliest time the start date is ignored.
+            end_date_time_series (datetime.datetime): The new end time of list of load profiles.
+                A zero demand entry is appended to series if the end date provided is later than
+                the last load profile entry in the list.
+            x_axis_time_period_timedelta (datetime.timedelta, optional): _description_. Defaults to datetime.timedelta(weeks=1).
+            resample_frequency (str, optional): _description_. Defaults to "1min".
+
+        Raises:
+            Exception: Raises an exception for ill defined start dates, end dates and x_axis_time_period_timedelta
+
+        Returns:
+            CarpetPlotMatrix: A dataclass that contains all information that is necessary to create
+            a load profile carpet plot.
+        """
         if not list_of_load_profile_entries:
-            print("Plot empty load profile")
-        number_of_periods = (
-            end_date_time_series - start_date_time_series
[…]
