### 2024-03-07 4b31647
Refactor load profile post processing, add tests, add documentation
