+++ src/ethos_penalps/post_processing/report_generator/carpet_plot_page.py
@@ -4,12 +4,12 @@ import datapane
 import matplotlib.pyplot
 import pandas
 
-from ethos_penalps.data_classes import LoadType
+from ethos_penalps.data_classes import LoadType, CarpetPlotMatrix, CarpetPlotMatrixEmpty
 from ethos_penalps.post_processing.load_profile_entry_post_processor import (
     LoadProfileEntryPostProcessor,
 )
-from ethos_penalps.post_processing.load_profile_post_processor import (
-    LoadProfilePostProcessor,
+from ethos_penalps.post_processing.carpet_plot_load_profile_generator import (
+    CarpetPlotLoadProfileGenerator,
 )
 from ethos_penalps.post_processing.network_analyzer import NetworkAnalyzer
 from ethos_penalps.post_processing.report_generator.report_options import (
@@ -38,7 +38,7 @@ class CarpetPlotPageGenerator:
 
             carpet_plot_list = []
             load_profile_matrix_dict_by_load_profile: dict[
-                LoadType, list[pandas.DataFrame]
+                str, list[CarpetPlotMatrix]
             ] = {}
 
             load_profile_entry_post_processor_for_start_time_check = (
@@ -69,9 +69,6 @@ class CarpetPlotPageGenerator:
                     period=report_generator_options.carpet_plot_options.x_axis_time_delta,
                     list_of_list_of_load_profile_entries=self.production_plan.load_profile_handler.get_list_of_list_of_all_load_profile_entries(),
                 )
-                start_date_for_display = (
-                    report_generator_options.carpet_plot_options.start_date
-                )
 
                 for (
                     stream_name,
@@ -88,7 +85,7 @@ class CarpetPlotPageGenerator:
                         load_type = stream_load_profile_collections.load_type_dict[
                             load_type_uuid
                         ]
-                        load_profile_post_processor = LoadProfilePostProcessor()
+                        load_profile_post_processor = CarpetPlotLoadProfileGenerator()
                         load_profile_df_matrix = load_profile_post_processor.convert_lpg_load_profile_to_data_frame_matrix(
                             list_of_load_profile_entries=list_of_load_profile_entries,
                             end_date_time_series=report_generator_options.carpet_plot_options.end_date,
@@ -97,46 +94,51 @@ class CarpetPlotPageGenerator:
                             resample_frequency=report_generator_options.carpet_plot_options.resample_frequency,
                             object_name=stream_name,
                         )
-                        fig = load_profile_post_processor.plot_load_profile_carpet_from_data_frame_matrix(
-                            carpet_plot_load_profile_matrix=load_profile_df_matrix,
-                            load_type_name=load_type.name,
-                        )
-                        caption = (
-                            "Stream name: "
-                            + str(stream_name)
-                            + "Load type: "
-                            + str(load_type.name)
-                        )
+                        if type(load_profile_df_matrix) is CarpetPlotMatrixEmpty:
+                            pass
+                        elif type(load_profile_df_matrix) is CarpetPlotMatrix:
+                            fig = load_profile_post_processor.plot_load_profile_carpet_from_data_frame_matrix(
+                                carpet_plot_load_profile_matrix=load_profile_df_matrix
+                            )
+                            caption = (
+                                "Stream name: "
+                                + str(stream_name)
+                                + "Load type: "
+                                + str(load_type.name)
+                            )
 
-                        file_name = (
-                            "stream_load_profile_"
-                            + str(stream_name)
-                            + "-"
-                            + str(load_type.name)
-                            + output_file_extension_with_dot
-                        )
-                        file_name = file_name.replace(" ", "_")
-                        output_file = os.path.join(
-                            load_profile_carpet_plot_directory_path, file_name
-                        )
-                        fig.savefig(
-                            output_file,
[…]
