### 2024-03-07 b65af29
Fix carpet plot
