+++ examples/tutorial/_1_cooking_example.py
@@ -1,6 +1,9 @@
 import datetime
 import logging
 
+from typeguard import install_import_hook
+
+install_import_hook("ethos_penalps")
 from ethos_penalps.data_classes import Commodity, LoadType
 from ethos_penalps.enterprise import Enterprise
 from ethos_penalps.order_generator import NOrderGenerator
