### 2024-03-07 3824dd4
Fix carpet plot
