+++ src/ethos_penalps/post_processing/load_profile_post_processor.py
@@ -1,499 +1,0 @@
-import datetime
-import math
-
-
-import matplotlib.dates
-import matplotlib.figure
-import matplotlib.pyplot
-import matplotlib.ticker
-import numpy
-import pandas
-import seaborn
-
-from dataclasses import dataclass
-from ethos_penalps.data_classes import (
-    LoadProfileDataFrameMetaInformation,
-    LoadProfileEntry,
-    LoadType,
-)
-from ethos_penalps.utilities.units import Units
-from ethos_penalps.post_processing.load_profile_entry_post_processor import (
-    LoadProfileEntryPostProcessor,
-)
-
-from ethos_penalps.utilities.units import Units
-from ethos_penalps.utilities.general_functions import (
-    check_if_date_1_is_before_date_2,
-    check_if_date_1_is_before_or_at_date_2,
-    create_subscript_string_matplotlib,
-)
-from ethos_penalps.utilities.logger_ethos_penalps import PeNALPSLogger
-
-itcs_logger = PeNALPSLogger.get_logger_without_handler()
-
-
-@dataclass
-class CarpetPlotMatrix:
-    data_frame: pandas.DataFrame
-    start_date_time_series: datetime.datetime
-    end_date_time_series: datetime.datetime
-    x_axis_time_period_timedelta: datetime.timedelta
-    resample_frequency: str = "1min"
-
-
-class LoadProfilePostProcessor:
-    def __init__(self) -> None:
-        self.load_type: LoadType
-        self.energy_unit: str
-        self.power_unit: str
-        self.start_time_of_input_load_profile: datetime.datetime
-        self.end_time_of_input_load_profile: datetime.datetime
-        self.processed_list_of_load_profile_entries: list[LoadProfileEntry]
-        self.start_date_resampled_load_profile: datetime.datetime
-        self.end_date_resampled_load_profile: datetime.datetime
-
-    def convert_lpg_load_profile_to_data_frame_matrix(
-        self,
-        list_of_load_profile_entries: list[LoadProfileEntry],
-        start_date_time_series: datetime.datetime,
-        end_date_time_series: datetime.datetime,
-        x_axis_time_period_timedelta: datetime.timedelta = datetime.timedelta(weeks=1),
-        resample_frequency: str = "1min",
-    ) -> CarpetPlotMatrix:
-        if not list_of_load_profile_entries:
-            print("Plot empty load profile")
-        number_of_periods = (
-            end_date_time_series - start_date_time_series
-        ) / x_axis_time_period_timedelta
-        if number_of_periods <= 0:
-            raise Exception(
-                "No positive number periods. Start time: "
-                + str(start_date_time_series)
-                + " End time: "
-                + str(end_date_time_series)
-            )
-
-        load_profile_entry_post_processor = LoadProfileEntryPostProcessor()
-        list_of_load_profile_entries = (
-            load_profile_entry_post_processor.homogenize_list_of_load_profiles_entries(
-                list_of_load_profile_entries=list_of_load_profile_entries,
-                start_date_time_series=start_date_time_series,
-                end_date_time_series=end_date_time_series,
-                resample_frequency=resample_frequency,
-            )
-        )
-
-        resampled_load_profile_data_frame = (
-            self.convert_list_of_load_profile_entries_to_data_frame(
-                list_of_load_profile_entries=list_of_load_profile_entries
-            )
-        )
-        load_profile_matrix = (
-            self.convert_time_series_to_power_matrix_with_period_columns(
-                input_data_frame=resampled_load_profile_data_frame,
-                periodic_time_delta=x_axis_time_period_timedelta,
-            )
-        )
-        carpet_plot_matrix = CarpetPlotMatrix(
[…]
