+++ examples/toffee_production/simulation_starter.py
@@ -111,7 +111,7 @@ enterprise.start_simulation()
 enterprise.create_post_simulation_report(
     gantt_chart_end_date=time_data.global_end_date,
     gantt_chart_start_date=time_data.global_end_date - datetime.timedelta(hours=28),
-    x_axis_time_delta=datetime.timedelta(hours=24),
+    x_axis_time_delta=datetime.timedelta(days=1),
     start_date=time_data.global_end_date - datetime.timedelta(days=365),
     end_date=time_data.global_end_date,
 )
