+++ src/ethos_penalps/post_processing/carpet_plot_load_profile_generator.py
@@ -1,0 +1,553 @@
+import datetime
+
+import matplotlib.dates
+import matplotlib.figure
+import matplotlib.pyplot
+import matplotlib.ticker
+import numpy
+import pandas
+import seaborn
+import warnings
+from ethos_penalps.data_classes import (
+    CarpetPlotMatrix,
+    CarpetPlotMatrixEmpty,
+    LoadProfileDataFrameMetaInformation,
+    LoadProfileEntry,
+    LoadType,
+    ListOfLoadProfileMetaData,
+    ListLoadProfileMetaDataEmpty,
+)
+from ethos_penalps.post_processing.load_profile_entry_post_processor import (
+    LoadProfileEntryPostProcessor,
+    ListOfLoadProfileEntryAnalyzer,
+)
+from ethos_penalps.utilities.exceptions_and_warnings import (
+    MisconfigurationError,
+    LoadProfileInconsistencyWarning,
+)
+from ethos_penalps.utilities.general_functions import (
+    check_if_date_1_is_before_date_2,
+    check_if_date_1_is_before_or_at_date_2,
+    create_subscript_string_matplotlib,
+)
+from ethos_penalps.utilities.logger_ethos_penalps import PeNALPSLogger
+from ethos_penalps.utilities.units import Units
+
+logger = PeNALPSLogger.get_logger_without_handler()
+
+
+class CarpetPlotLoadProfileGenerator(LoadProfileEntryPostProcessor):
+    """This class is used to create carpet plots from a
+    list of load profile entries.
+    """
+
+    def __init__(self) -> None:
+        pass
+
+    def convert_lpg_load_profile_to_carpet_plot(
+        self,
+        list_of_load_profile_entries: list[LoadProfileEntry],
+        start_date: datetime.datetime,
+        end_date: datetime.datetime,
+        object_name: str,
+        x_axis_time_period_timedelta: datetime.timedelta = datetime.timedelta(weeks=1),
+        resample_frequency: str = "1min",
+    ) -> matplotlib.figure.Figure | None:
+        carpet_plot_load_profile_matrix = (
+            self.convert_lpg_load_profile_to_data_frame_matrix(
+                list_of_load_profile_entries=list_of_load_profile_entries,
+                start_date_time_series=start_date,
+                end_date_time_series=end_date,
+                x_axis_time_period_timedelta=x_axis_time_period_timedelta,
+                resample_frequency=resample_frequency,
+                object_name=object_name,
+            )
+        )
+        if type(carpet_plot_load_profile_matrix) is CarpetPlotMatrixEmpty:
+            logger.debug("No load profile to plot for object: %s", object_name)
+            figure = None
+        elif type(carpet_plot_load_profile_matrix) is CarpetPlotMatrix:
+            figure = self.plot_load_profile_carpet_from_data_frame_matrix(
+                carpet_plot_load_profile_matrix=carpet_plot_load_profile_matrix,
+            )
+        else:
+            raise Exception("Unexpected datatype")
+        return figure
+
+    def convert_lpg_load_profile_to_data_frame_matrix(
+        self,
+        list_of_load_profile_entries: list[LoadProfileEntry],
+        start_date_time_series: datetime.datetime,
+        end_date_time_series: datetime.datetime,
+        object_name: str,
+        x_axis_time_period_timedelta: datetime.timedelta = datetime.timedelta(weeks=1),
+        resample_frequency: str = "1min",
+    ) -> CarpetPlotMatrix | CarpetPlotMatrixEmpty:
+        """Converts a list of load profiles into a CarpetPlotMatrix
+        which is used to create carpet plots.
+            1. The list of load profiles is homogenized
+            2. The list of load profiles is converted into a data frame
[…]
