### 2024-03-07 df7e48d
Change checkout commit in test pipeline of conda-forge
