+++ test/post_processing/test_compress_power_load_profile_meta_data.py
@@ -1,0 +1,208 @@
+import datetime
+import itertools
+import os
+import pathlib
+from dataclasses import dataclass
+
+import pint
+from ethos_penalps.data_classes import (
+    ListLoadProfileMetaDataEmpty,
+    ListOfLoadProfileMetaData,
+    CarpetPlotMatrix,
+    LoadProfileEntry,
+    LoadType,
+)
+from ethos_penalps.post_processing.load_profile_entry_post_processor import (
+    ListOfLoadProfileEntryAnalyzer,
+    LoadProfileEntryPostProcessor,
+)
+from ethos_penalps.post_processing.carpet_plot_load_profile_generator import (
+    CarpetPlotMatrix,
+    LoadProfileEntryPostProcessor,
+    CarpetPlotLoadProfileGenerator,
+)
+from ethos_penalps.utilities.units import Units
+
+
+@dataclass
+class EnergyData:
+    list_of_energy_values: list[float]
+    energy_unit: str
+    time_step: datetime.timedelta
+    load_type: LoadType
+
+    def __post_init__(self):
+        self.energy_value_cycle_operator: itertools.cycle = itertools.cycle(
+            self.list_of_energy_values
+        )
+
+    def reset_cycle(self):
+        self.energy_value_cycle_operator: itertools.cycle = itertools.cycle(
+            self.list_of_energy_values
+        )
+
+    def convert_to_power_data(self, target_power_unit: str) -> "PowerData":
+        list_of_power_values = []
+        for energy_value in self.list_of_energy_values:
+            power_value = Units.convert_energy_to_power(
+                energy_value=energy_value,
+                energy_unit=self.energy_unit,
+                time_step=self.time_step,
+                target_power_unit=target_power_unit,
+            )
+            list_of_power_values.append(power_value)
+        power_data = PowerData(
+            list_of_power_values=list_of_power_values,
+            power_unit=target_power_unit,
+            time_step=self.time_step,
+            load_type=self.load_type,
+        )
+        return power_data
+
+
+@dataclass
+class PowerData:
+    list_of_power_values: list[float]
+    power_unit: str
+    time_step: datetime.timedelta
+    load_type: LoadType
+
+    def __post_init__(self):
+        self.power_value_cycle_operator: itertools.cycle = itertools.cycle(
+            self.list_of_power_values
+        )
+
+    def reset_cycle(self):
+        self.power_value_cycle_operator: itertools.cycle = itertools.cycle(
+            self.list_of_power_values
+        )
+
+    def convert_to_energy_data(self, target_energy_unit: str) -> "EnergyData":
+        list_of_energy_values = []
+        for power_value in self.list_of_power_values:
+            energy_value = Units.convert_power_to_energy(
+                power_value=power_value,
+                power_unit=self.power_unit,
+                time_step=self.time_step,
+                target_energy_unit=target_energy_unit,
+            )
+            list_of_energy_values.append(energy_value)
+        energy_data = EnergyData(
+            list_of_energy_values=list_of_energy_values,
+            energy_unit=target_energy_unit,
+            time_step=self.time_step,
+            load_type=self.load_type,
+        )
+        return energy_data
+
+
+def create_sample_list_of_load_profile_entry_data(
+    energy_data: EnergyData,
+    power_data: PowerData,
+    total_start_time: datetime.datetime,
+    total_end_time: datetime.datetime,
+):
+    current_start_time = total_end_time - energy_data.time_step
+    current_end_time = total_end_time
+    list_of_load_profile_entries = []
+    number_of_time_steps = (total_end_time - total_start_time) / energy_data.time_step
+    for time_step in range(int(number_of_time_steps)):
[…]
