### 2024-03-08 cafdd86
Refactor post processing
