+++ test/production_plan_tests/test_production_plan_restoration.py
@@ -1,10 +1,12 @@
 from ethos_penalps.production_plan import ProductionPlan
-from ethos_penalps.load_profile_calculator import LoadProfileHandler
+from ethos_penalps.load_profile_calculator import LoadProfileHandlerSimulation
 import pathlib
 
 
 def test_load_production_plan():
-    production_plan = ProductionPlan(load_profile_handler=LoadProfileHandler())
+    production_plan = ProductionPlan(
+        load_profile_handler=LoadProfileHandlerSimulation()
+    )
     parent_directory = pathlib.Path(__file__).parent.parent.absolute()
     case_folder_name = r"cases"
     case_1_name = r"case_1"
