+++ test/post_processing/test_carpet_plot/test_carpet_plot.py
@@ -2,280 +2,119 @@ import datetime
 import itertools
 import os
 import pathlib
+from collections.abc import Iterable
 from dataclasses import dataclass
 
 import pint
+import pytest
+from ethos_penalps.utilities.general_functions import (
+    convert_date_time_to_string,
+    ResultPathGenerator,
+)
 from ethos_penalps.data_classes import LoadProfileEntry, LoadType
-from ethos_penalps.post_processing.carpet_plot_load_profile_generator import (
-    CarpetPlotMatrix,
+from ethos_penalps.post_processing.tikz_visualizations.carpet_plot_load_profile_generator import (
     CarpetPlotLoadProfileGenerator,
+    CarpetPlotMatrix,
 )
 from ethos_penalps.utilities.units import Units
 
 
-@dataclass
-class EnergyData:
-    list_of_energy_values: list[float]
-    energy_unit: str
-    time_step: datetime.timedelta
-    load_type: LoadType
+from test.post_processing.load_profile_generator_for_tests import (
+    EnergyData,
+    PowerData,
+    BaseLoadProfileGeneratorClass,
+)
 
-    def __post_init__(self):
-        self.energy_value_cycle_operator: itertools.cycle = itertools.cycle(
-            self.list_of_energy_values
-        )
+pytestmark = pytest.mark.carpet_plot_tests
 
-    def reset_cycle(self):
-        self.energy_value_cycle_operator: itertools.cycle = itertools.cycle(
-            self.list_of_energy_values
-        )
 
-    def convert_to_power_data(self, target_power_unit: str) -> "PowerData":
-        list_of_power_values = []
-        for energy_value in self.list_of_energy_values:
-            power_value = Units.convert_energy_to_power(
-                energy_value=energy_value,
-                energy_unit=self.energy_unit,
-                time_step=self.time_step,
-                target_power_unit=target_power_unit,
+class TestCarpetPlot(BaseLoadProfileGeneratorClass):
+
+    @pytest.mark.parametrize(
+        ("x_axis_time_period_timedelta"),
+        [
+            (datetime.timedelta(hours=1)),
+            (datetime.timedelta(days=1)),
+            (datetime.timedelta(weeks=1)),
+        ],
+    )
+    @pytest.mark.parametrize(
+        ("time_step"),
+        [
+            (datetime.timedelta(minutes=5)),
+            (datetime.timedelta(minutes=30)),
+            (datetime.timedelta(minutes=60)),
+            (datetime.timedelta(minutes=90)),
+        ],
+    )
+    @pytest.mark.parametrize(
+        (
+            "total_start_time",
+            "total_end_time",
+        ),
+        [
+            (
+                datetime.datetime(year=2022, month=1, day=1),
+                datetime.datetime(year=2022, month=12, day=31),
             )
-            list_of_power_values.append(power_value)
-        power_data = PowerData(
-            list_of_power_values=list_of_power_values,
-            power_unit=target_power_unit,
-            time_step=self.time_step,
-            load_type=self.load_type,
+        ],
+    )
+    @pytest.mark.parametrize(
+        ("resample_frequency"),
+        [("5min"), ("30min"), ("1h")],
+    )
+    def test_carpet_plot_creation(
+        self,
+        x_axis_time_period_timedelta: datetime.timedelta,
+        total_start_time: datetime.datetime,
+        total_end_time: datetime.datetime,
+        time_step: datetime.timedelta,
+        resample_frequency: str,
+    ):
+
+        list_of_load_profile_entries = self.create_load_profile_entry_list(
+            total_start_time=total_start_time,
+            total_end_time=total_end_time,
+            time_step=time_step,
         )
-        return power_data
-
-
-@dataclass
-class PowerData:
[…]

[…]
