+++ src/ethos_penalps/post_processing/tikz_visualizations/carpet_plot_load_profile_generator.py
@@ -9,6 +9,7 @@ import numpy
 import pandas
 import pint
 import seaborn
+import math
 
 from ethos_penalps.data_classes import (
     CarpetPlotMatrix,
@@ -82,6 +83,7 @@ class CarpetPlotLoadProfileGenerator(LoadProfileEntryPostProcessor):
         start_date_time_series: datetime.datetime,
         end_date_time_series: datetime.datetime,
         object_name: str,
+        object_type: str,
         x_axis_time_period_timedelta: datetime.timedelta = datetime.timedelta(weeks=1),
         resample_frequency: str = "1min",
     ) -> CarpetPlotMatrix | CarpetPlotMatrixEmpty:
@@ -128,31 +130,43 @@ class CarpetPlotLoadProfileGenerator(LoadProfileEntryPostProcessor):
                     be positive."""
                 )
 
-            list_of_load_profile_meta_data = (
-                self.homogenize_list_of_load_profiles_entries(
-                    list_of_load_profile_entries=list_of_load_profile_entries,
+            list_of_load_profile_meta_data = self.create_load_profile_meta_data(
+                list_of_load_profile_entries=list_of_load_profile_entries,
+                start_date_time_series=start_date_time_series,
+                end_date_time_series=end_date_time_series,
+                object_name=object_name,
+                object_type=object_type,
+            )
+            assert type(list_of_load_profile_meta_data) is LoadProfileMetaData
+            list_of_load_profile_meta_data_resampled = (
+                self.resample_load_profile_meta_data(
+                    load_profile_meta_data=list_of_load_profile_meta_data,
+                    start_date=start_date_time_series,
+                    end_date=end_date_time_series,
+                    x_axis_time_period_timedelta=x_axis_time_period_timedelta,
+                    resample_frequency=resample_frequency,
+                )
+            )
+
+            carpet_plot_matrix: CarpetPlotMatrix | CarpetPlotMatrixEmpty
+            if (
+                type(list_of_load_profile_meta_data_resampled)
+                is LoadProfileMetaDataResampled
+            ):
+
+                carpet_plot_matrix = self.convert_load_profile_meta_data_to_carpet_plot_matrix(
+                    load_profile_meta_data_resampled=list_of_load_profile_meta_data_resampled,
+                    x_axis_period_time_delta=x_axis_time_period_timedelta,
                     start_date_time_series=start_date_time_series,
                     end_date_time_series=end_date_time_series,
                     resample_frequency=resample_frequency,
                     object_name=object_name,
                 )
-            )
 
-            carpet_plot_matrix: CarpetPlotMatrix | CarpetPlotMatrixEmpty
-            if type(list_of_load_profile_meta_data) is LoadProfileMetaDataResampled:
-
-                carpet_plot_matrix = (
-                    self.convert_load_profile_meta_data_to_carpet_plot_matrix(
-                        list_of_load_profile_meta_data=list_of_load_profile_meta_data,
-                        x_axis_period_time_delta=x_axis_time_period_timedelta,
-                        start_date_time_series=start_date_time_series,
-                        end_date_time_series=end_date_time_series,
-                        resample_frequency=resample_frequency,
-                        object_name=object_name,
-                    )
-                )
-
-            elif type(list_of_load_profile_meta_data) is EmptyLoadProfileMetadata:
+            elif (
+                type(list_of_load_profile_meta_data_resampled)
+                is EmptyLoadProfileMetadata
+            ):
                 logger.debug(
[…]
