+++ test/storage/storage.py
@@ -47,4 +47,4 @@ storage = Storage(
     state_data_container=ProductionProcessStateContainer(),
 )
 storage.create_all_storage_production_plan_entry()
-print("")
+
