### 2024-03-12 #6358 @julian-belina
@Extraweich Still thank you for your effort. The problem is that the standard conda environment solver is very slow when trying to solve complex enviroments. There is another option to speed up the installation proceedure using anaconda.  You can install the libmama solver in you conda base environement. This is the solver that is also used by mamba.

`
conda install -n base conda-libmamba-solver
`
`
conda config --set solver libmamba
`

Unfortunately, if you have installed a lot of things into you base environment, this installation of the solver itself can take forever. I will add this remark to my installation proceedure.
