+++ test/post_processing/test_load_profile_post_processing.py
@@ -16,7 +16,7 @@ from ethos_penalps.data_classes import (
 from ethos_penalps.post_processing.load_profile_entry_post_processor import (
     LoadProfileEntryPostProcessor,
 )
-from ethos_penalps.post_processing.tikz_visualizations.carpet_plot_load_profile_generator import (
+from ethos_penalps.post_processing.time_series_visualizations.carpet_plot_load_profile_generator import (
     CarpetPlotLoadProfileGenerator,
 )
 from ethos_penalps.utilities.exceptions_and_warnings import (

[…]
