+++ pyproject.toml
@@ -49,7 +49,7 @@ filterwarnings = [
     "ignore: Deprecated call to `pkg_resources.declare_namespace",
     "ignore: distutils Version classes are deprecated.",
     "ignore: pkg_resources is deprecated as an API.",
-    "ignore: Deprecated API features detected! These feature(s) are not compatible with SQLAlchemy 2.0"
+    "ignore: Deprecated API features detected!"
     
 
 ] # How to configure Filterwarning:minversion = "6.0"
