+++ .github/workflows/test.yml
@@ -44,17 +44,17 @@ jobs:
             echo "LS Done"
             mamba install -n test_env ethos_penalps pytest pytest-xdist
             echo "Installation done"
-            conda list -n test_env        
+            conda list -n test_env -e      
         - name: get_penalps_version
           shell: pwsh
           id: penalps-version
           # use ethos-penalps instead of ethos_penalps for looup due to diffrent formatting in conda list output
           run: |
-              $PENALPS_STRING = conda list -e -n test_env ethos-penalps | Out-String
+              $PENALPS_STRING = conda list -n test_env ethos_penalps -e | Out-String
               echo "The penalps string is"
               echo  $PENALPS_STRING
               echo "Start to extract the version"
-              $PENALPS_VERSION = [regex]::Match($PENALPS_STRING, 'ethos-penalps=(\d+\.\d+\.\d+)=').Groups[1].Value 
+              $PENALPS_VERSION = [regex]::Match($PENALPS_STRING, 'ethos_penalps=(\d+\.\d+\.\d+)=').Groups[1].Value 
               "PENALPS_VERSION=$PENALPS_VERSION" >> $GITHUB_OUTPUT
               echo "Print the extracted version"
               echo $PENALPS_VERSION
