+++ .github/workflows/test.yml
@@ -45,25 +45,16 @@ jobs:
             mamba install -n test_env ethos_penalps pytest pytest-xdist
             echo "Installation done"
             conda list -n test_env -e
-            $PENALPS_STRING = conda list -n test_env ethos-penalps -e | Out-String
-            echo "The penalps string is"
-            echo $PENALPS_STRING
-            echo "Start to extract the version"
-            $PENALPS_VERSION = [regex]::Match($PENALPS_STRING, 'ethos-penalps-penalpslps=(\d+\.\d+\.\d+)=').Groups[1].Value 
-            "PENALPS_VERSION=$PENALPS_VERSION" >> $GITHUB_OUTPUT
-            echo "Print the extracted version"
-            echo $PENALPS_VERSION
-            echo "Terminate get penalps version"
         - name: get_penalps_version
           shell: pwsh
           id: penalps-version
           # use ethos-penalps instead of ethos_penalps for looup due to diffrent formatting in conda list output
           run: |
-              $PENALPS_STRING = conda list -n test_env ethos-penalps -e | Out-String
+              $PENALPS_STRING = conda list -n test_env ethos_penalps -e | Out-String
               echo "The penalps string is"
               echo $PENALPS_STRING
               echo "Start to extract the version"
-              $PENALPS_VERSION = [regex]::Match($PENALPS_STRING, 'ethos-penalps-penalpslps=(\d+\.\d+\.\d+)=').Groups[1].Value 
+              $PENALPS_VERSION = [regex]::Match($PENALPS_STRING, 'ethos_penalps=(\d+\.\d+\.\d+)=').Groups[1].Value 
               "PENALPS_VERSION=$PENALPS_VERSION" >> $GITHUB_OUTPUT
               echo "Print the extracted version"
               echo $PENALPS_VERSION
@@ -140,7 +131,7 @@ jobs:
             cd basic_examples
             conda run --name penalps_env python batch_to_batch_1_node_example.py
             cd ..
-            cd B-Pillar_manufacturing
+            cd b_pillar_manufacturing
             conda run --name penalps_env python simulation_starter.py
             echo "Running examples terminated"
 
