+++ src/ethos_penalps/post_processing/process_summary.py
@@ -1,514 +1,0 @@
-import matplotlib.pyplot
-import pandas
-
-from ethos_penalps.data_classes import (
-    LoadType,
-    ProcessStateEnergyLoadData,
-    ProcessStateEnergyLoadDataBasedOnStreamMass,
-    ProductEnergyData,
-    StreamLoadEnergyData,
-)
-from ethos_penalps.load_profile_calculator import (
-    LoadProfileHandlerSimulation,
-    StreamSpecificEnergyDataHandler,
-)
-from ethos_penalps.node_operations import ProductionOrder
-from ethos_penalps.process_nodes.process_node import ProcessNode
-from ethos_penalps.process_nodes.process_step import ProcessStep
-from ethos_penalps.process_nodes.sink import Sink
-from ethos_penalps.process_nodes.source import Source
-from ethos_penalps.production_plan import ProductionPlan
-from ethos_penalps.stream import (
-    BatchStream,
-    BatchStreamProductionPlanEntry,
-    ContinuousStream,
-    ContinuousStreamProductionPlanEntry,
-    ProcessStepProductionPlanEntry,
-    StreamDataFrameMetaInformation,
-)
-from ethos_penalps.stream_handler import StreamHandler
-from ethos_penalps.utilities.general_functions import (
-    create_subscript_string_matplotlib,
-    get_sub,
-    get_super,
-)
-from ethos_penalps.utilities.units import Units
-
-
-class ProcessOverViewGenerator:
-    def __init__(
-        self,
-        process_node_dict: dict[str, ProcessNode],
-        load_profile_handler: LoadProfileHandlerSimulation,
-        stream_handler: StreamHandler,
-        order_dictionary: dict[float, ProductionOrder],
-        enterprise_name: str,
-        production_plan: ProductionPlan,
-    ) -> None:
-        self.process_node_dict: dict[str, ProcessNode] = process_node_dict
-        self.load_profile_handler: LoadProfileHandlerSimulation = load_profile_handler
-        self.stream_handler: StreamHandler = stream_handler
-        self.load_type_dict_for_product_specific_energy_demand: dict[
-            LoadType, ProductEnergyData
-        ] = {}
-        self.order_dictionary: dict[float, ProductionOrder] = order_dictionary
-        self.stream_mass_to_product_mass_dict: dict[str, float] = {}
-        self.enterprise_name: str = enterprise_name
-        self.production_plan: ProductionPlan = production_plan
-        self.stream_specific_energy_demand_product_corrected: dict[
-            str, dict[LoadType, float]
-        ] = {}
-
-    def determine_total_energy_demand_per_ton_of_end_product(self):
-        self.create_empty_product_energy_dict()
-        self.determine_stream_mass_to_product_mass_conversion_factors()
-        for (
-            stream_name
-        ) in (
-            self.load_profile_handler.stream_energy_data_collection.stream_energy_data_dict
-        ):
-            stream_to_product_energy_conversion_factor = (
-                self.stream_mass_to_product_mass_dict[stream_name]
-            )
-            self.add_stream_energy_data_to_product_energy_dict(
-                product_to_stream_conversion_factor=stream_to_product_energy_conversion_factor,
-                stream_name=stream_name,
-            )
-        for (
-            process_step_name
-        ) in self.load_profile_handler.process_step_energy_data_handler_dict:
-            self.add_process_step_energy_data_to_product_energy_dict(
-                process_step_name=process_step_name
-            )
-
-    def determine_stream_mass_to_product_mass_conversion_factors(self):
-        sink = self.get_sink()
-        current_stream_name = sink.get_input_stream_name()
-        current_conversion_factor = 1
-        self.stream_mass_to_product_mass_dict[current_stream_name] = (
-            current_conversion_factor
[…]
