+++ paper/paper.md
@@ -1,0 +1,122 @@
+---
+title: 'ETHOS.PeNALPS: A Tool for the Load Profile Simulation of Industrial Processes Based on a Material Flow Simulation'
+tags:
+  - Python
+  - Load Profile
+  - Industry
+  - Manufacturing
+  - Energy Simulation
+  - Industrial Production
+  - Materials Processing
+authors:
+  - name: Julian Belina
+    orcid: 0000-0002-5878-2936
+    affiliation: "1, 2"
+  - name: Noah Pflugradt
+    orcid: 0000-0002-1982-8794
+    affiliation: "1, 2"
+  - name: Detlef Stolten
+    orcid: 0000-0002-1671-3262
+    affiliation: "1, 2, 3"
+affiliations:
+ - name: Jülich Aachen Research Alliance, JARA-Energy, Jülich, Aachen, Germany
+   index: 1
+ - name: Forschungszentrum Jülich GmbH, Institute of Energy and Climate Research – Techno-economic Systems Analysis (IEK-3), 52425 Jülich, Germany
+   index: 2
+ - name: RWTH Aachen University, Chair for Fuel Cells, Faculty of Mechanical Engineering, 52062 Aachen, Germany
+   index: 3
+date: 30 September 2023
+bibliography: citavi.bib
+---
+
+
+# Summary
+ETHOS.PeNALPS (Petri Net Agent-based Load Profile Simulator) is a Python library for
+simulating the load profiles of industrial manufacturing processes for arbitrary energy carriers. It is part of [ETHOS (Energy Transformation Pathway Optimization Suite)](https://go.fzj.de/ethos_suite). Load profiles are energy demand time series. Processes that can be simulated using ETHOS.PeNALPS include, for example, steel, paper, and industrial food production.
+
+Figure \ref{fig:Main Component Overview} shows the main conceptual objects of ETHOS.PeNALPS which are: 
+
+- Generic model objects
+- Material flow simulations 
+- Production plans
+- Result load profiles
+
+The model of the material flow simulation is created by users based on generic simulation objects. After the material flow simulation is completed, a set of production orders is passed to the model to start the simulation. The simulation generates a production plan that tracks the activity of each node to fulfill the requested set of orders. Based on the activity in the production plan, the load profiles are created for each node in therein.
+
+![The main components of ETHOS.PeNALPS are the generic models objects, material flow simulation, production plan and load profiles.\label{fig:Main Component Overview}](main_component_overview.png){width=100%}
+
+# Statement of Need
+Load profiles are of particular interest for assessing energy demand fluctuations in energy system modeling and design. For industrial processes, load profiles are often not available for open research due to:
+
+- Efforts by enterprises to protect commercial secrets;
+- Missing measurements; 
+-	Unstructured energy data collection in enterprises;
+-	Novelty of the industrial processes and their currently missing implementation.
+
+ETHOS.PeNALPS can provide these missing load profiles via simulation.
+
+# Method
+
+ETHOS.PeNALPS is capable of modeling noncyclical industrial production networks. The simulation is created from generic objects, which are shown in Figure \ref{fig:Main Component Overview}. The most important components are the generic nodes that handle and create the material requests as agents. The generic node types are: 
+
+- Source
+- Sink
+- Process step
+- Storage
+
+These nodes are connected by streams that determine the direction of the material flow in the simulation. Sequentially-dependent nodes and streams are combined in so-called process chains. Multiple or single process chains are integrated into a network level. Multiple chains in a single network level model the parallel operation of similar equipment. Multiple network levels can be used to model network features of the industrial process. 
+
[…]
