### 2024-03-13 #6358 @julian-belina
> Functional issue: I opened a ticket about the energy plots that appear to sometimes use blank date in one of the included examples. See https://github.com/FZJ-IEK3-VSA/ETHOS_PeNALPS/issues/3

@stoldtjo The carpet plots have been reworked. They are included in the 1.0.3 version. The conda forge package will be published in the next hours.

@dpzhuX Sorry for the late reply. I ended up refactoring parts of my post processing which took longer than expectet.

> 1. Validation of examples: There are two examples mentioned in the package documentation. Is there any validation available for these examples? Additionally, I noticed that the results for these examples are not displayed on the documentation website

Currently no validated examples are available. The examples are just showcases for the modelling capabilities and i added a comment to the examples which explicitly states this. I am going to create a validation example for my thesis.

> 2. Load profiles: How were the load profiles (c) generated in Figure 2? After running the toffee_production example, I saw the simulation report but did not find load profiles similar to those depicted in load profiles (c) of Figure 2.

The figure of the load profiles c is created manually. The report provides carpet plots which are suited better to display long load profiles. I chose the other depiction in figure c to show the dependence of the load profile to the production plan.

>3. Parallel chain design: For the parallel chain design in toffee_production example, I observed that the primary distinction between chain 1 and chain 2 seems to be their names (i.e., Machine 1 vs Machine 2). If we want to have numerous parallel chains (e.g., 1000 chains) with identical designs, is it necessary to manually duplicate and rename each file for every chain? If this is the case, I think the code design might need to improve.

This is possible. It is demonstrated in the tutorial _5_connect_four_process_steps.

>4. Multiple sources and sinks: A general question for this package that I did not find answer in the documentation. The documentation does not explicitly address whether a chain can include multiple sources and sinks for a complex topology. Can you confirm if the package can accommodate such a configuration:

ETHOS.PeNALPS can not model a production system with the topology of your example. Currently it is only possible to model a single sink and source per network level. I have added an [article which explains the possible topologies of the production systems](https://ethospenalps.readthedocs.io/en/latest/ethos_penalps_articles/model_topology.html).

>5. Issue with running an example: I have an issue when attempting to run the _4_connect_four_process_steps from the tutorial folder, I have the following error:

Thank you for pointing out the error. Somehow the import path got altered. I added the examples to my test pipline of version 1.0.3 to prevent such errors in future.

@Extraweich I have added the remarks concerning Anaconda to the Readme and installation section in the documentation.
