### 2024-03-13 FZJ-IEK3-VSA/ETHOS_PeNALPS#3 @julian-belina
_
![image](https://github.com/FZJ-IEK3-VSA/ETHOS_PeNALPS/assets/56728940/b339fa48-46d9-4d48-ba3a-6bc3410f8624)

I have updated the carpet plots. The dates on the x-axis are now shown properly.
