### 2024-03-13 #6358 @srmnitc
> I have tried the installation several times and still have not found a way to get it to work. When I use Anaconda, I get stuck in solving environments. Installing Mamba, as requested in the installation instructions, is not an option for me at the moment as it is not meant to be installed alongside Anaconda and I cannot compromise on the environments I am currently using. I'm not sure if this is just a problem with my system or if it's down to my own human incompetence. With two other reviewers already providing valuable comments who are far more versed in this area than I am, I assume that my current role as a reviewer adds no value to the process. @srmnitc, I would therefore politely request to be removed as a reviewer so as not to further impede the process.

Thank you once again for agreeing to be the reviewer and your feedback. I will of course unassign you from here. Thanks a lot once again!
