### 2024-03-14 #6358 @stoldtjo
@srmnitc Thanks I will go through the checklist again now but I don't see significant changes in response to the criticism and suggestions on the paper I stated. Please let me know if I am overextending what is expected from the paper for the JOSS but otherwise I find that to be a crucial point.
