### 2024-03-14 FZJ-IEK3-VSA/ETHOS_PeNALPS#6 @julian-belina
Improve API documentation
> API Docs: I usually work a lot with Microsoft docs and I used to work with Plant Simulation references. Compared to either the API Docs are lacking. Without checking the example, doing some elaborate guess work or actually looking into the project's code I would feel somewhat lost which methods and attributes are responsible for which purpose and what I should expect from them in terms of functionality. ethos_penalps.process_nodes.process_step, for instance, is a little better than the majority of samples I checked (e.g., ethos_penalps.simulation_data.container_simulation_data).

The API documentation of core objects needs to be improved as suggested by stoldjo in the  [JOSS review](https://github.com/openjournals/joss-reviews/issues/6358)
