### 2024-03-14 #6358 @julian-belina
@stoldtjo @srmnitc  I am aware that the API documentation still needs improvement. I have created an issue to keep track of the progress.

@stoldtjo The refactoring of the post processing took more time than expected and I still need time to read the literature, which seems verry helpfull so far. Thanks for the suggestion!

I will try to improve the statement of need and structure of the paper. The original idea of the software was to provide energy demand data for indurial sectors for energy system modeling. The industrial energy demand data that is currently used in many studies does not consider any fluctuations within a year at all. So any fluctuations would be an improvement to the state of the art. It is difficult to determine a sufficient level of detail for the results because the temporal resolution of the energy system optimization models keeps increasing. Many models run at a resolution of one hour but that might change. I will make a suggestion when I am done with the literature.
