### 2024-03-14 FZJ-IEK3-VSA/ETHOS_PeNALPS#7 @julian-belina
Improve Statement of Need
The statement of need requires improvement. The target audience needs to be adressed more clearly
[as pointed out in the review](https://github.com/openjournals/joss-reviews/issues/6358)

> Paper Section "Statement of Need" is weak because what has been noted as a requirement is kept very generic and general. Particularly compared with the implemented energy model the software is capable of serving the need. The implemented energy consumption model (average consumption per object state) is suitable for a small set of potential analysis questions but many more would hardly profit from this modeling approach. See also this chapter: http://dx.doi.org/10.1007/978-3-031-34218-9_1
