### 2024-03-14 FZJ-IEK3-VSA/ETHOS_PeNALPS#8 @julian-belina
Improve Other Tools and Methods Section
The Other Tools and Methods section does not show the plethora of orther Methods and Tools. It should reference another review and select relevant work.  [This was pointed out in the JOSS review](https://github.com/openjournals/joss-reviews/issues/6358)

> Paper Section "Other Tools and Methods" is weak and fails to give a solid overview of other tools and methods. A number of review papers have been published and they provide some additional pointers on relevant works and publications. Particularly early works in the field from between 2005 and 2015 contained a number of works that report on non-commercial tools developed by various research group. Authors Wohlgemuth, Pawletta, Andersson, Grundmeier, Heilala, Rippel, Sobottka come to mind. It might suffice to reference appropriate review papers but it definitely seems advisable that previous works are checked for their licensing status to ensure the statement "most of these are not open source" is not just driven by those works which are based on commercial tools.

> Paper List of References: The selection of references feels rather arbitrary to someone who has been following this field of research for some years. The authors should either attempt to present a more complete picture of existing research/publications or refer to publications providing it and commenting why specific examples were selected and how they compare to other non-cited publications (e.g., to elaborate a specific example in a little more depth while mentioning that the example stands for a group of similar approaches).
