+++ paper/citavi.bib
@@ -1,4 +1,4 @@
-﻿% This file was created with Citavi 6.17.0.0
+﻿% This file was created with Citavi 6.18.0.1
 
 @misc{.ChocoTecCarastar,
  author = {{Chocotech GmbH}},
@@ -207,6 +207,33 @@
  journal = {ACM Computing Surveys},
  doi = {10.1145/356698.356702},
  file = {3-540-65306-6{\_}14:Attachments/3-540-65306-6{\_}14.pdf:application/pdf;356698.356702:Attachments/356698.356702.pdf:application/pdf}
+}
+
+
+@article{Priesmann.2021,
+ abstract = {The analysis of energy scenarios for future energy systems requires appropriate data. However, while more or less detailed data on energy production is often available, appropriate data on energy consumption is often scarce. In our JERICHO-E-usage dataset, we provide comprehensive data on useful energy consumption patterns for heat, cold,~mechanical energy,~information and communication, and light in high spatial and temporal resolution. Furthermore, we distinguish between residential, industrial, commerce, and mobility consumers. For our dataset, we aggregate bottom-up data and disaggregate top-down data both to the NUTS2 level. The NUTS2 level serves as an interface to validate our combined method approach and the calculations. We combine a multitude of data sources such as weather time series, standard load profiles, census data, movement data, and employment figures to increase the scope, validity, and reproducibility for energy system modeling. The focus of our JERICHO-E-usage dataset on useful energy consumption might be of particular interest to researchers who analyze energy scenarios where renewable electricity is largely substituted for fossil fuel (sector coupling).},
+ author = {Priesmann, Jan and Nolting, Lars and Kockel, Christina and Praktiknjo, Aaron},
+ year = {2021},
+ title = {Time series of useful energy consumption patterns for energy system modeling},
+ pages = {148},
+ volume = {8},
+ number = {1},
+ journal = {Scientific data},
+ doi = {10.1038/s41597-021-00907-w},
+ file = {Priesmann, Nolting et al. 2021 - Time series of useful energy:Attachments/Priesmann, Nolting et al. 2021 - Time series of useful energy.pdf:application/pdf}
+}
+
+
+@article{Prina.2020,
+ author = {Prina, Matteo Giacomo and Manzolini, Giampaolo and Moser, David and Nastasi, Benedetto and Sparber, Wolfram},
+ year = {2020},
+ title = {Classification and challenges of bottom-up energy system models - A review},
+ pages = {109917},
+ volume = {129},
+ issn = {13640321},
+ journal = {Renewable and Sustainable Energy Reviews},
+ doi = {10.1016/j.rser.2020.109917},
+ file = {1-s2.0-S1364032120302082-main:Attachments/1-s2.0-S1364032120302082-main.pdf:application/pdf}
 }
 
 
