+++ paper/paper.md
@@ -29,55 +29,60 @@ date: 30 September 2023
 bibliography: citavi.bib
 ---
 
+# Summary 
 
-# Summary
-ETHOS.PeNALPS (Petri Net Agent-based Load Profile Simulator) is a Python library for
-simulating the load profiles of industrial manufacturing processes for arbitrary energy carriers. It is part of [ETHOS (Energy Transformation Pathway Optimization Suite)](https://go.fzj.de/ethos_suite). Load profiles are energy demand time series. It can model the fluctuations of industrial processes that are caused logistic dependencies of machines and state based activity of machines. If the model Processes that can be simulated using ETHOS.PeNALPS include, for example, steel, paper, and industrial food production. If the load profile can be modelled in a sufficient detail depends on the process, the available input data and the required temporal resolution.
+ETHOS.PeNALPS (Petri Net Agent-based Load Profile Simulator) is a Python library designed to simulate the load profiles of industrial manufacturing processes for arbitrary energy carriers. It is part of the [ETHOS (Energy Transformation Pathway Optimization Suite)](https://go.fzj.de/ethos_suite). Load profiles
+are time series of energy demand. The library models the material flow of industrial processes and the activity of individual machines during production. ETHOS.PeNALPS is capable of simulating processes such as steel, paper, and industrial food production. ETHOS.PeNALPS is able to model non-cyclic industrial production networks. 
 
-Figure \ref{fig:Main Component Overview} shows the main conceptual objects of ETHOS.PeNALPS which are: 
+Figure \ref{fig:Main Component Overview} shows the main conceptual objects of ETHOS.PeNALPS, which are: 
 
 - Generic model objects
 - Material flow simulations 
 - Production plans
 - Result load profiles
 
-The model of the material flow simulation is created by users based on generic simulation objects. After the material flow simulation is completed, a set of production orders is passed to the model to start the simulation. The simulation generates a production plan that tracks the activity of each node to fulfill the requested set of orders. Based on the activity in the production plan, the load profiles are created for each node in therein.
+The user creates the process model based on generic simulation objects. Once the user completes process model, the model receives a set of production orders to initiate the simulation. The simulation generates a production plan that tracks the activity of each node to fulfill the requested set of orders. Load profiles are then created based on the activity in the production plan. The process steps's load profiles are modeled using a Petri Net with an extensible number of states that determine their activity and energy demand during production. 
 
-![The main components of ETHOS.PeNALPS are the generic models objects, material flow simulation, production plan and load profiles.\label{fig:Main Component Overview}](main_component_overview.png){width=100%}
+![The main components of ETHOS.PeNALPS are the generic model objects, material flow simulation, production plan and load profiles.\label{fig:Main Component Overview}](main_component_overview.png){width=100%}
+
 
 # Statement of Need
-Long term energy system models are an important tool to investigate how real energy systems could and should develop. However, creating long term energy system models with a high spatial resolution, temporal resolution at a high level of data and model transparency remains a challenge.~\cite[S.~12]{Prina.2020} One of the challenges is to obtain suitable energy demand data for the industrial sector for the relevant time horizon. Some data for industrial sectors is available for e.g. for the german industrial sector in 2015 ~\cite[S.~5--6]{Priesmann.2021}. But the the industrial sector and thus its load profiles is going to change to due to decarbonization efforts. Also for a lot of regions such data is missing entirely. One approach to overcome this lack of data is to model the load profile of the industrial sector bottom-up from the load profiles of processes that constitute it. Load profiles of these processes are often not available due load profiles are often not available for open research due to:
 
-- Efforts by enterprises to protect commercial secrets;
[…]
