+++ paper/citavi.bib
@@ -1,5 +1,23 @@
 ﻿% This file was created with Citavi 6.18.0.1
 
+@proceedings{.2012,
+ year = {2012},
+ title = {Proceedings Title: Proceedings of the 2012 Winter Simulation Conference (WSC)},
+ publisher = {IEEE},
+ isbn = {978-1-4673-4782-2}
+}
+
+
+@proceedings{.2012b,
+ year = {2012},
+ title = {Proceedings of the 2012 Symposium on Emerging Applications of M{\&}S in Industry and Academia Symposium},
+ address = {San Diego, CA, USA},
+ publisher = {{Society for Computer Simulation International}},
+ isbn = {9781618397874},
+ series = {EAIA '12}
+}
+
+
 @misc{.ChocoTecCarastar,
  author = {{Chocotech GmbH}},
  year = {24.10.2023},
@@ -7,6 +25,34 @@
  url = {https://www.chocotech.de/technologie/kochen/carastarr},
  urldate = {24.10.2023},
  file = {CARASTAR{\circledR} 24102023:Attachments/CARASTAR{\circledR} 24102023.pdf:application/pdf}
+}
+
+
+@inproceedings{Andersson.2012,
+ author = {Andersson, Jon and Skoogh, Anders and Johansson, Bjorn},
+ title = {Evaluation of methods used for life-cycle assessments in Discrete Event Simulation},
+ pages = {1--12},
+ publisher = {IEEE},
+ isbn = {978-1-4673-4782-2},
+ booktitle = {Proceedings Title: Proceedings of the 2012 Winter Simulation Conference (WSC)},
+ year = {2012},
+ doi = {10.1109/WSC.2012.6465094},
+ file = {Evaluation{\_}of{\_}methods{\_}used{\_}for{\_}life-cycle{\_}assessments{\_}in{\_}Discrete{\_}Event{\_}Simulation:Attachments/Evaluation{\_}of{\_}methods{\_}used{\_}for{\_}life-cycle{\_}assessments{\_}in{\_}Discrete{\_}Event{\_}Simulation.pdf:application/pdf}
+}
+
+
+@inproceedings{Andersson.2012b,
+ abstract = {Ecolabled products have shown a competitive advantage to other products. Regulatory changes and market pressure creates an increased need for environmental impact assessments. The dominating method for environmental impact assessments - life cycle assessment (LCA) lacks support to properly analyze the dynamic aspects of business operations and production processes. This Paper proposes to use discrete event simulation to support more extensive and detailed environmental assessments on selected parts of the production process, keeping simplicity for parts of less importance and interest.},
+ author = {Andersson, Jon and Johansson, Bj{\"o}rn and Berglund, Jonatan and Skoogh, Anders},
+ title = {Framework for ecolabeling using discrete event simulation},
+ keywords = {discrete event simulation;ecolabeling;environmental assessment tool},
+ publisher = {{Society for Computer Simulation International}},
+ isbn = {9781618397874},
+ series = {EAIA '12},
+ booktitle = {Proceedings of the 2012 Symposium on Emerging Applications of M{\&}S in Industry and Academia Symposium},
+ year = {2012},
+ address = {San Diego, CA, USA},
+ file = {local{\_}156301:Attachments/local{\_}156301.pdf:application/pdf}
 }
 
 
@@ -49,6 +95,17 @@
 }
 
 
+@misc{DassaultSystemes.2023,
+ abstract = {Explore DELMIA - digital manufacturing and planning software which helps industries to collaborate, model and optimize their operations.},
+ author = {{Dassault Syst{\`e}mes}},
+ year = {2023},
+ title = {DELMIA},
[…]
