+++ documentation/ethos_penalps_articles/contribution_guide.md
@@ -1,0 +1,48 @@
+# Contributions Guide
+
+Contributions are welcome, and they are greatly appreciated! Every little bit
+helps, and credit will always be given.
+
+You can contribute in many ways:
+
+## Types of Contributions
+
+
+### Report Bugs
+
+
+Report bugs at https://github.com/FZJ-IEK3-VSA/ETHOS_PeNALPS/issues.
+
+If you are reporting a bug, please include:
+
+* Your operating system name and version.
+* Any details about your local setup that might be helpful in troubleshooting.
+* Detailed steps to reproduce the bug.
+
+### Fix Bugs
+
+Look through the GitHub issues for bugs. Anything tagged with "bug" and "help
+wanted" is open to whoever wants to implement it.
+
+### Implement Features
+
+Look through the GitHub issues for features. Anything tagged with "enhancement"
+and "help wanted" is open to whoever wants to implement it.
+
+### Write Documentation
+
+ETHOS.PeNALPS could always use more documentation, whether as part of the
+official ETHOS.PeNALPS docs, in docstrings, or even on the web in blog posts,
+articles, and such.
+
+### Submit Feedback
+
+
+The best way to send feedback is to file an issue at https://github.com/FZJ-IEK3-VSA/ETHOS_PeNALPS/issues.
+
+If you are proposing a feature:
+
+- Explain in detail how it would work.
+- Keep the scope as narrow as possible, to make it easier to implement.
+- Remember that this is a volunteer-driven project, and that contributions
+  are welcome :)
