+++ documentation/_toc.yml
@@ -12,6 +12,7 @@ parts:
     maxdepth: 3
     chapters:
       - file: ethos_penalps_tutorial/0_overview.md
+      - file: ethos_penalps_tutorial/0_1_installation.md
       - file: ethos_penalps_tutorial/1_single_cooker_process_chain.md
       - file: ethos_penalps_tutorial/2_add_more_states.md
       - file: ethos_penalps_tutorial/3_add_more_cookers_for_parallel_operations.md
@@ -22,6 +23,8 @@ parts:
     chapters:
       - file: ethos_penalps_articles/simulation_workflow.md
       - file: ethos_penalps_articles/model_topology.md
+      - file: ethos_penalps_articles/roadmap.md
+      - file: ethos_penalps_articles/contribution_guide.md
 
   - caption: API
     chapters:

[…]
