### 2024-03-22 16b8b1d
Update paper and API documentation
