+++ paper/paper.md
@@ -41,7 +41,7 @@ Figure \ref{fig:Main Component Overview} shows the main conceptual objects of ET
 - Production plans
 - Result load profiles
 
-The user creates the process model based on generic simulation objects. Once the user completes process model, the model receives a set of production orders to initiate the simulation. The simulation generates a production plan that tracks the activity of each node to fulfill the requested set of orders. Load profiles are then created based on the activity in the production plan. The process steps's load profiles are modeled using a Petri Net with an extensible number of states that determine their activity and energy demand during production. 
+The user creates the process model based on generic simulation objects. Once the user completes process model, the model receives a set of production orders to initiate the simulation. The simulation generates a production plan that tracks the activity of each node to fulfill the requested set of orders. Load profiles are then created based on the activity in the production plan. The process steps's load profiles are modeled using a Petri net with an extensible number of states that determine their activity and energy demand during production. 
 
 ![The main components of ETHOS.PeNALPS are the generic model objects, material flow simulation, production plan and load profiles.\label{fig:Main Component Overview}](main_component_overview.png){width=100%}
 
@@ -55,90 +55,48 @@ Energy system models are tools that provide guidance on future energy systems [@
 -	Unstructured collection of energy data in companies;
 -	Novelty of the industrial processes and their current lack of implementation.
 
-ETHOS.PeNALPS can support the creation of an energy system model by providing load profiles for industrial processes. While many industrial processes and their load profiles have been previously simulated, most have not published load profiles and simulation implementation under an open source license. This creates a research gap, despite similar work having already been done.
+ETHOS.PeNALPS can support the creation of an energy system model by providing load profiles for industrial processes. While many industrial processes and their load profiles have been previously simulated, most have not published load profiles and simulation implementation under an open-source license. This creates a research gap, despite similar work having already been done.
 
-The suitability of load profiles simulated by ETHOS.PeNALPS for an energy system model depends on the available input data, the type of process being modeled, and the required temporal resolution. These three aspects are interdependent, making it impossible to make a static selection of industrial processes that can be provided for all energy system models. At lower temporal resolutions, effects may occur that cannot be modeled under the assumptions of a deterministic Petri Net of machine state and average energy consumption per state. Additionally, the temporal resolution of energy system models is evolving. In 2020, a temporal resolution of one hour was considered high for long-term energy system models [@Prina.2020 p.10]. Currently, studies may require load profiles with a resolution as low as one minute [xxx]. 
+The suitability of load profiles simulated by ETHOS.PeNALPS for an energy system model depends on the available input data, the type of process being modeled, and the required temporal resolutions. These three aspects are interdependent, making it impossible to make a static selection of industrial processes that can be provided for all energy system models. At lower temporal resolutions, effects may occur that cannot be modeled under the assumptions of a deterministic Petri net of machine state and average energy consumption per state. Additionally, the temporal resolution of energy system models is evolving. In 2020, a temporal resolution of one hour was considered high for long-term energy system models [@Prina.2020 p.10]. Currently, studies may require load profiles with a resolution as low as one minute [@Omoyele.2024 p.12-13]. 
 
 # Method
-The simulation is generated from generic objects, as shown in Figure \ref{fig:Main Component Overview}. The main components are the generic nodes that create and manage material requests as agents. These nodes include: 
+
[…]
