+++ paper/convert_to_word.py
@@ -1,3 +1,0 @@
-import pypandoc
-
-pypandoc.co
