+++ src/ethos_penalps/petri_net/__init__.py
@@ -1,0 +1,2 @@
+"""Contains all Modules and Classes for the Petri net. Rearranging is still in Progress.
+"""
