+++ src/ethos_penalps/post_processing/report_generator/__init__.py
@@ -1,0 +1,2 @@
+"""Contains all modules and classes that are required for the generation of reports.
+"""

[…]
