+++ src/ethos_penalps/process_step_data.py
@@ -17,6 +17,8 @@ from ethos_penalps.time_data import TimeData
 
 
 class ProcessStepData:
+    """Contains all objects that required to alter the state of the discrete event simulation."""
+
     def __init__(
         self,
         process_step_name: str,
@@ -34,6 +36,11 @@ class ProcessStepData:
         self.load_profile_handler: LoadProfileHandlerSimulation = load_profile_handler
 
     def restore_time_data(self, new_time_data: TimeData):
+        """Resets the time data to a previous state.
+
+        Args:
+            new_time_data (TimeData): TimeData which contains the new state.
+        """
         self.time_data.last_idle_time = new_time_data.last_idle_time
         self.time_data.last_process_state_switch_time = (
             new_time_data.last_process_state_switch_time
@@ -50,7 +57,5 @@ class ProcessStepData:
         self.time_data.storage_last_update_time = new_time_data.storage_last_update_time
 
     def validate_input_stream(self):
-        # # Input stream must be added to storage level before it is shifted to validated stream list
-        # self.main_mass_balance.storage.add_validated_input_stream_to_storage_level()
-        # Shift input stream to validated stream list
+        """Input stream must be added to storage level before it is shifted to validated stream list"""
         self.state_data_container.validate_input_stream()
