+++ src/ethos_penalps/post_processing/time_series_visualizations/carpet_plot_load_profile_generator.py
@@ -47,36 +47,6 @@ class CarpetPlotLoadProfileGenerator(LoadProfileEntryPostProcessor):
     def __init__(self) -> None:
         pass
 
-    def convert_lpg_load_profile_to_carpet_plot(
-        self,
-        list_of_load_profile_entries: list[LoadProfileEntry],
-        start_date: datetime.datetime,
-        end_date: datetime.datetime,
-        object_name: str,
-        x_axis_time_period_timedelta: datetime.timedelta = datetime.timedelta(weeks=1),
-        resample_frequency: str = "1min",
-    ) -> matplotlib.figure.Figure | None:
-        carpet_plot_load_profile_matrix = (
-            self.convert_lpg_load_profile_to_data_frame_matrix(
-                list_of_load_profile_entries=list_of_load_profile_entries,
-                start_date_time_series=start_date,
-                end_date_time_series=end_date,
-                x_axis_time_period_timedelta=x_axis_time_period_timedelta,
-                resample_frequency=resample_frequency,
-                object_name=object_name,
-            )
-        )
-        if type(carpet_plot_load_profile_matrix) is CarpetPlotMatrixEmpty:
-            logger.debug("No load profile to plot for object: %s", object_name)
-            figure = None
-        elif type(carpet_plot_load_profile_matrix) is CarpetPlotMatrix:
-            figure = self.plot_load_profile_carpet_from_data_frame_matrix(
-                carpet_plot_load_profile_matrix=carpet_plot_load_profile_matrix,
-            )
-        else:
-            raise Exception("Unexpected datatype")
-        return figure
-
     def convert_lpg_load_profile_to_data_frame_matrix(
         self,
         list_of_load_profile_entries: list[LoadProfileEntry],
@@ -181,6 +151,14 @@ class CarpetPlotLoadProfileGenerator(LoadProfileEntryPostProcessor):
         return carpet_plot_matrix
 
     def invert_list(self, list_to_invert: list):
+        """Inverts the input list.
+
+        Args:
+            list_to_invert (list): List to be inverted.
+
+        Returns:
+            _type_: Inverted list.
+        """
         return list_to_invert[::-1]
 
     def convert_load_profile_meta_data_to_carpet_plot_matrix(
@@ -302,6 +280,15 @@ class CarpetPlotLoadProfileGenerator(LoadProfileEntryPostProcessor):
     def get_energy_amount_from_data_frame(
         self, load_profile_data_frame: pandas.DataFrame
     ) -> float:
+        """Return the total energy from the data frame.
+
+        Args:
+            load_profile_data_frame (pandas.DataFrame): Data frame that
+                is created from a list of LoadProfileEntry.
+
+        Returns:
+            float: Value of the total energy amount.
+        """
         energy_amount = load_profile_data_frame["energy_quantity"].sum()
         return energy_amount
 
@@ -309,6 +296,19 @@ class CarpetPlotLoadProfileGenerator(LoadProfileEntryPostProcessor):
         self,
         carpet_plot_matrix: CarpetPlotMatrix,
     ) -> float:
+        """Returns the energy amount of the CarpetPlotMatrix
+        that is calculated from the power entries.
+
+        Args:
+            carpet_plot_matrix (CarpetPlotMatrix): Contains the power
+                values which should be plotted using a carpet plot.
+                Also contains additional meta information about the power
+                values.
+
+
+        Returns:
+            float: Total energy amount.
+        """
         carpet_plot_matrix_data_frame = carpet_plot_matrix.data_frame
[…]
