+++ src/ethos_penalps/data_classes.py
@@ -71,6 +71,11 @@ class EmptyMetaDataInformation:
 
 @dataclass()
 class ProductEnergyData(DataClassJsonMixin):
+    """Summarizes the specific energy demand that is
+    required for an end product. It considers previous
+    energy demand and conversion factors from previous
+    steps."""
+
     product_commodity: Commodity
     specific_energy_demand: float
     load_type: LoadType
@@ -80,6 +85,10 @@ class ProductEnergyData(DataClassJsonMixin):
 
 @dataclass(frozen=True, eq=True, unsafe_hash=True)
 class StreamLoadEnergyData(DataClassJsonMixin):
+    """Summarizes the energy data that is required
+    to converts a stream state to a LoadProfileEntry.
+    """
+
     stream_name: str
     specific_energy_demand: float
     load_type: LoadType
@@ -89,6 +98,11 @@ class StreamLoadEnergyData(DataClassJsonMixin):
 
 @dataclass(slots=True, frozen=True)
 class LoadProfileEntry(DataClassJsonMixin):
+    """Represents the energy demand of
+    stream or process step in the given time
+    period.
+    """
+
     load_type: LoadType
     start_time: datetime.datetime
     end_time: datetime.datetime
@@ -100,6 +114,16 @@ class LoadProfileEntry(DataClassJsonMixin):
     def _adjust_power_unit(
         self, new_power_value: float, new_power_unit: str
     ) -> "LoadProfileEntry":
+        """Converts the power quantity to a
+        new value and unit.
+
+        Args:
+            new_power_value (float): New power value.
+            new_power_unit (str): New power unit string.
+
+        Returns:
+            LoadProfileEntry: Converted load profile entry.
+        """
         load_profile_entry = LoadProfileEntry(
             load_type=self.load_type,
             start_time=self.start_time,
@@ -113,11 +137,19 @@ class LoadProfileEntry(DataClassJsonMixin):
 
 
 class LoopCounter:
+    """Counts the iterations within a ProcessChain
+    during the simulation.
+    """
+
     loop_number: float | str = "Loop has not started"
 
 
 @dataclass(kw_only=True, frozen=True, slots=True)
 class ProcessStepProductionPlanEntry(DataClassJsonMixin):
+    """Summarizes the activity of a process step in
+    a discrete time period.
+    """
+
     process_step_name: str
     process_state_name: str
     start_time: datetime.datetime
@@ -128,6 +160,10 @@ class ProcessStepProductionPlanEntry(DataClassJsonMixin):
 
 @dataclass(frozen=True, slots=True)
 class StorageProductionPlanEntry(DataClassJsonMixin):
+    """Represents the storage level in a discrete
+    time period.
+    """
+
     process_step_name: str
     start_time: datetime.datetime
     end_time: datetime.datetime
@@ -139,12 +175,21 @@ class StorageProductionPlanEntry(DataClassJsonMixin):
 
 @dataclass(frozen=True, eq=True, slots=True)
 class StateConnector:
+    """Represents the connection between
+    two states of the petri net in a ProcessStep.
+    """
+
     start_state_name: str
     end_state_name: str
 
 
 @dataclass
 class ProcessStepDataFrameMetaInformation:
+    """Provides the simulation results of a
+    ProcessStep and additional meta information
+    about this data and the.
+    """
+
     data_frame: pandas.DataFrame
     process_step_name: str
     list_of_process_state_names: list[str]
@@ -154,6 +199,11 @@ class ProcessStepDataFrameMetaInformation:
 
 @dataclass
 class StorageDataFrameMetaInformation:
+    """Provides the simulation results of
+    a stream and additional meta information
+    about this data and the.
+    """
+
     data_frame: pandas.DataFrame
     process_step_name: str
     commodity: Commodity
[…]
