+++ paper/paper.md
@@ -31,8 +31,7 @@ bibliography: citavi.bib
 
 # Summary 
 
-ETHOS.PeNALPS (Petri Net Agent-based Load Profile Simulator) is a Python library designed to simulate the load profiles of industrial manufacturing processes for arbitrary energy carriers. It is part of the [ETHOS (Energy Transformation Pathway Optimization Suite)](https://go.fzj.de/ethos_suite). Load profiles
-are time series of energy demand. The library models the material flow of industrial processes and the activity of individual machines during production. ETHOS.PeNALPS is capable of simulating processes such as steel, paper, and industrial food production. ETHOS.PeNALPS is able to model non-cyclic industrial production networks. 
+ETHOS.PeNALPS (Petri Net Agent-based Load Profile Simulator) is a Python library designed to simulate the load profiles of industrial manufacturing processes for arbitrary energy carriers. It is part of the [ETHOS (Energy Transformation Pathway Optimization Suite)](https://go.fzj.de/ethos_suite). Load profiles are time series of energy demand. The library models the material flow of industrial processes and the activity of individual machines during production. ETHOS.PeNALPS is capable of simulating processes such as steel, paper, and industrial food production. ETHOS.PeNALPS is able to model non-cyclic industrial production networks. 
 
 Figure \ref{fig:Main Component Overview} shows the main conceptual objects of ETHOS.PeNALPS, which are: 
 
@@ -48,16 +47,16 @@ The user creates the process model based on generic simulation objects. Once the
 
 # Statement of Need
 
-Energy system models are tools that provide guidance on future energy systems [@Prina.2020 p.1]. However, building long-term models with high spatial and temporal resolution and transparent input data remains a challenge [@Prina.2020 p.12].  For instance, historical load profiles for the German industrial sector in 2015 are available [@Priesmann.2021 p.5-6], while load profiles for other regions are not currently available. Furthermore, decarbonization efforts will cause changes in the industrial sector, creating a need for load profiles of future scenarios. To address the lack of sectoral load profiles for the industry, Boßmann and Stafell [-@Bomann.2015 p.1321] demonstrated the use of a bottom-up approach to create load profiles. Therefore, it is necessary to obtain load profiles of the industrial processes that are part of the industrial sector. However, these profiles are often unavailable for open research due to:
+Energy system models are tools that provide guidance on future energy systems, which are currently undergoing significant changes to due global efforts to reduce dependence on fossil fuels. [@Prina.2020 p.1]. However, building long-term models with high spatial and temporal resolution and transparent input data remains a challenge [@Prina.2020 p.12].  For instance, historical load profiles for the German industrial sector in 2015 are available [@Priesmann.2021 p.5-6], while load profiles for other regions are not currently available. Furthermore, decarbonization efforts will cause changes in the industrial sector, creating a need for load profiles of future scenarios. To address the lack of sectoral load profiles for the industry, Boßmann and Stafell [-@Bomann.2015 p.1321] demonstrated the use of a bottom-up approach. Therefore, it is necessary to obtain load profiles of the industrial processes that are part of the industrial sector. However, these profiles are often unavailable for open research due to:
 
 - Companies' efforts to protect commercial secrets;
 - Missing measurements; 
 -	Unstructured collection of energy data in companies;
 -	Novelty of the industrial processes and their current lack of implementation.
 
-ETHOS.PeNALPS can support the creation of an energy system model by providing load profiles for industrial processes. While many industrial processes and their load profiles have been previously simulated, most have not published load profiles and simulation implementation under an open-source license. This creates a research gap, despite similar work having already been done.
[…]
