+++ src/ethos_penalps/stream_node_distributor.py
@@ -26,6 +26,10 @@ from ethos_penalps.utilities.general_functions import (
 
 @dataclass
 class SplittedOrderCollection:
+    """Represents a set of orders that has been splitted
+    among multiple ProcessChains
+    """
+
     stream_name: str
     commodity: Commodity
     process_chain_identifier: ProcessChainIdentifier
@@ -34,6 +38,11 @@ class SplittedOrderCollection:
     current_order_number: int = 0
 
     def check_if_order_are_empty(self):
+        """Raises an error if an order is empty.
+
+        Raises:
+            MisconfigurationError: Is raised if the order is empty.
+        """
         if self.order_data_frame.empty:
             raise MisconfigurationError(
                 "A splitted order of chain: "
@@ -42,6 +51,14 @@ class SplittedOrderCollection:
             )
 
     def get_order_by_order_number(self, order_number: int) -> ProductionOrder:
+        """Returns an order based on the order number.
+
+        Args:
+            order_number (int): Number that identifies the order.
+
+        Returns:
+            ProductionOrder: Order for a product oder intermediate product.
+        """
         order_data_frame_row = self.order_data_frame.iloc[order_number]
         production_order = ProductionOrder(
             production_target=order_data_frame_row.loc["production_target"],
@@ -56,6 +73,12 @@ class SplittedOrderCollection:
         return production_order
 
     def update_order(self, produced_mass: numbers.Number):
+        """Updates the mass that is already produced.
+
+        Args:
+            produced_mass (numbers.Number): The additional mass that has been
+                produced and should be added to the order.
+        """
         self.order_data_frame.at[self.current_order_number, "produced_mass"] = (
             self.order_data_frame.at[self.current_order_number, "produced_mass"]
             + produced_mass
@@ -63,12 +86,23 @@ class SplittedOrderCollection:
 
 
 class OrderDistributor:
+    """Distributes the orders among multiple ProcessChains."""
+
     def __init__(
         self,
         stream_handler: StreamHandler,
         production_order_collection: OrderCollection,
         node_name: str,
     ) -> None:
+        """
+
+        Args:
+            stream_handler (StreamHandler): Container for all streams
+                connected the sink to the process chains.
+            production_order_collection (OrderCollection): Set of orders
+                that should be delivered to the sink.
+            node_name (str): Name of the sink that should receive the orders.
+        """
         self.node_name: str = node_name
         self.dict_of_stream_names: dict[ProcessChainIdentifier, str] = {}
         self.stream_handler: StreamHandler = stream_handler
@@ -80,12 +114,24 @@ class OrderDistributor:
         self.current_splitted_order: SplittedOrderCollection
 
     def update_order_collection(self, new_order_collection: OrderCollection):
+        """Adds new orders to the current set of orders.
+
+        Args:
+            new_order_collection (OrderCollection): The set of orders
+                that should be added to the current one.
+        """
         self.order_collection.append_order_collection(new_order_collection)
         self.split_production_order_dict()
 
     def set_current_splitted_order_by_chain_identifier(
         self, process_chain_identifier: ProcessChainIdentifier
     ):
+        """Activates the set of splitted orders for a new process chain.
+
+        Args:
+            process_chain_identifier (ProcessChainIdentifier): The process chain
+                that should that is simulated next.
+        """
         self.current_splitted_order = self.dict_of_splitted_order[
             process_chain_identifier
         ]
@@ -93,9 +139,20 @@ class OrderDistributor:
     def get_current_splitted_order(
         self,
     ) -> SplittedOrderCollection:
+        """Returns the set of orders that is currently simulated.
+
+        Returns:
+            SplittedOrderCollection: Current set of orders.
+        """
         return self.current_splitted_order
 
[…]
