### 2024-03-26 8e2b7a3
Update paper and API Documentation. Release 1.0.4
