+++ paper/paper.md
@@ -31,7 +31,7 @@ bibliography: citavi.bib
 
 # Summary 
 
-ETHOS.PeNALPS (Petri Net Agent-based Load Profile Simulator) is a Python library designed to simulate the load profiles of industrial manufacturing processes for arbitrary energy carriers. It is part of the [ETHOS (Energy Transformation Pathway Optimization Suite)](https://go.fzj.de/ethos_suite). Load profiles are time series of energy demand. The library models the material flow of industrial processes and the activity of individual machines during production. ETHOS.PeNALPS is capable of simulating processes such as steel, paper, and industrial food production. ETHOS.PeNALPS is able to model non-cyclic industrial production networks. 
+ETHOS.PeNALPS (Petri Net Agent-based Load Profile Simulator) is a Python library designed to simulate the load profiles of industrial manufacturing processes for arbitrary energy carriers. It is part of the [ETHOS (Energy Transformation Pathway Optimization Suite)](https://go.fzj.de/ethos_suite). Load profiles are time series of energy demand. The library models the material flow of industrial processes and the activity of individual machines during production. ETHOS.PeNALPS is capable of simulating processes such as steel, paper, and industrial food production. ETHOS.PeNALPS can model non-cyclic industrial production networks. 
 
 Figure \ref{fig:Main Component Overview} shows the main conceptual objects of ETHOS.PeNALPS, which are: 
 
@@ -40,7 +40,7 @@ Figure \ref{fig:Main Component Overview} shows the main conceptual objects of ET
 - Production plans
 - Result load profiles
 
-The user creates the process model based on generic simulation objects. Once the user completes process model, the model receives a set of production orders to initiate the simulation. The simulation generates a production plan that tracks the activity of each node to fulfill the requested set of orders. Load profiles are then created based on the activity in the production plan. The process steps's load profiles are modeled using a Petri net with an extensible number of states that determine their activity and energy demand during production. 
+The user creates the process model based on generic simulation objects. Once the user completes the process model, the model receives a set of production orders to initiate the simulation. The simulation generates a production plan that tracks the activity of each node to fulfill the requested set of orders. Load profiles are then created based on the activity in the production plan. The process steps’ load profiles are modeled using a Petri net with an extensible number of states that determine their activity and energy demand during production. 
 
 ![The main components of ETHOS.PeNALPS are the generic model objects, material flow simulation, production plan and load profiles.\label{fig:Main Component Overview}](main_component_overview.png){width=100%}
 
@@ -56,13 +56,13 @@ Energy system models are tools that provide guidance on future energy systems, w
 
 ETHOS.PeNALPS can support the creation of an energy system model by providing load profiles for industrial processes. While many industrial processes and their load profiles have been previously simulated, most have not published load profiles and simulation implementations under an open-source license. This creates a research gap, despite similar work having already been done.
 
[…]
