+++ src/ethos_penalps/process_state_switch.py
@@ -25,20 +25,34 @@ class ProcessStateSwitch(ABC):
 
     def calculate_next_event_time_backward(self) -> datetime.datetime:
         """Determines the switch time to the next state.
+
         Returns:
-            datetime.datetime: Time when a switch occurs from the 
+            datetime.datetime: Time when a switch occurs from the
                 end state to the start state.
         """
         raise NotImplementedError
 
 
 class ProcessStateSwitchDelay(ProcessStateSwitch):
+    """This delay is used to model a state switch after a static delay. This
+    is mainly used to model the length of an intermediate state.
+    """
+
     def __init__(
         self,
         process_step_data: ProcessStepData,
         state_connector: StateConnector,
         delay: datetime.timedelta,
     ) -> None:
+        """_summary_
+
+        Args:
+            process_step_data (ProcessStepData): Contains all data
+                that define the simulation state of the process step and all
+                methods to alter the state.
+            state_connector (StateConnector): Summarizes the connection of two process states.
+            delay (datetime.timedelta): The static delay after the process state switch occurs.
+        """
         if not isinstance(state_connector.start_state_name, str):
             raise Exception(
                 "A name of typ string should be supplied for process state identification"
@@ -72,11 +86,22 @@ class ProcessStateSwitchDelay(ProcessStateSwitch):
 
 
 class ProcessStateSwitchAtOutputStreamProvided(ProcessStateSwitch):
+    """This process state switch triggers at the start time of the output stream state.
+    The output state must be defined at the end state of this switch."""
+
     def __init__(
         self,
         process_step_data: ProcessStepData,
         state_connector: StateConnector,
     ) -> None:
+        """
+        Args:
+            process_step_data (ProcessStepData): Contains all data
+                that define the simulation state of the process step and all
+                methods to alter the state.
+            state_connector (StateConnector): Connects two states. The end state of
+                this switch should be an output stream providing state.
+        """
         if not isinstance(state_connector.start_state_name, str):
             raise Exception(
                 "A name of typ string should be supplied for process state identification"
@@ -95,8 +120,12 @@ class ProcessStateSwitchAtOutputStreamProvided(ProcessStateSwitch):
     def calculate_next_event_time_backward(
         self,
     ) -> datetime.datetime:
-        """This method calculates the time until the next storage event occurs
-        for each storage of the process step
+        """This method calculates the process state switch time. It triggers the switch
+        at the start time of the output stream state.
+
+        Returns:
+            datetime.datetime: The new switch time at the start time of the
+                output stream state.
         """
         if not isinstance(self.state_connector.start_state_name, str):
             raise Exception(
@@ -124,11 +153,23 @@ class ProcessStateSwitchAtOutputStreamProvided(ProcessStateSwitch):
 
 
 class ProcessStateSwitchAtInputStreamProvided(ProcessStateSwitch):
+    """This switch triggers at the the start time of the input stream state.
+    It must be connected to an input stream providing state as target state.
+    """
+
     def __init__(
         self,
         process_step_data: ProcessStepData,
         state_connector: StateConnector,
     ) -> None:
+        """
+        Args:
+            process_step_data (ProcessStepData): Contains all data
+                that define the simulation state of the process step and all
+                methods to alter the state.
+            state_connector (StateConnector): Connects two states. The end state of
+                this switch should be an input stream providing state.
+        """
         if not isinstance(state_connector.start_state_name, str):
             raise Exception(
                 "A name of typ string should be supplied for process state identification"
@@ -147,8 +188,12 @@ class ProcessStateSwitchAtInputStreamProvided(ProcessStateSwitch):
     def calculate_next_event_time_backward(
         self,
     ) -> datetime.datetime:
-        """This method calculates the time until the next storage event occurs
[…]
