+++ src/ethos_penalps/post_processing/tikz_visualizations/enterprise_graph_builder.py
@@ -108,12 +108,16 @@ class TikzNameHandler:
 
 @dataclass
 class EmptyTikzNode:
+    """Represents an empty tikz node which is not displayed"""
+
     unique_name: str = field(default_factory=get_new_uuid)
     tikz_options: str = "InfinitesimalNode"
 
 
 @dataclass
 class ProcessStepNode:
+    """Represents the a process step node"""
+
     tex_name: str
     name_in_simulation: str
     name_to_display: str
@@ -121,6 +125,8 @@ class ProcessStepNode:
 
 @dataclass
 class IntermediateStorage:
+    """Represents a storage which connects NetworkLevels"""
+
     sink: Sink
     source: Source
     unique_name: str
@@ -128,10 +134,18 @@ class IntermediateStorage:
 
 @dataclass(kw_only=True)
 class SinkRow:
+    """Represents a row in the tikz matrix which contains the sink."""
+
     main_sink: Sink
     unique_identification_name: str
 
     def create_tikz_string(self):
+        """Creates the string that represents the SinkRow in the
+        tex document.
+
+        Returns:
+            _type_: _description_
+        """
         sink_node = TikzNode(
             name_to_display=self.main_sink.name,
             unique_identification_name=self.unique_identification_name,
@@ -143,6 +157,13 @@ class SinkRow:
 
 @dataclass(kw_only=True)
 class SourceRow:
+    """Object that represents the source row
+    of the material flow system.
+
+    Returns:
+        _type_: _description_
+    """
+
     main_source: Source
     unique_identification_name: str
     unique_name_of_object_below_source: str
@@ -162,6 +183,14 @@ class SourceRow:
 
 @dataclass(kw_only=True)
 class SourceAndSinkRow(SinkRow, SourceRow):
+    """Object that represents the NetworkLevel Storage
+    of the material flow simulation.
+
+    Args:
+        SinkRow (_type_): _description_
+        SourceRow (_type_): _description_
+    """
+
     def __post_init__(self):
         self.name_to_display: str = self.main_sink.name
 
@@ -192,6 +221,7 @@ class SortedProcessChainLevel:
     unique_sink_name: str
 
     def __post_init__(self):
+        """Converts the input data into a structured form."""
         self.unique_chain_name = (
             self.tikz_name_handler.create_unique_tikz_identification_name(
                 input_name=self.process_chain.process_chain_identifier.chain_name
@@ -232,6 +262,8 @@ class SortedProcessChainLevel:
 
 @dataclass(kw_only=True)
 class SortedNetworkLevel:
+    """Sorted representation of the network level"""
+
     network_level: NetworkLevel
     tikz_name_handler: TikzNameHandler
     previous_source_row: SourceRow | None
@@ -294,6 +326,8 @@ class SortedNetworkLevel:
 
 @dataclass(kw_only=True)
 class FilledProcessStepChain:
+    """Representation of the filled process chain."""
+
     name_to_display: str
     maximum_chain_length: int
     unique_chain_name: str
@@ -312,6 +346,12 @@ class FilledProcessStepChain:
 
 @dataclass
 class NetworkLevelMatrix:
+    """Representation of a NetworkLevel for tikz plotting.
+
+    Returns:
+        _type_: _description_
+    """
+
     unique_tikz_name: str
     sorted_network_level: SortedNetworkLevel
 
@@ -350,11 +390,20 @@ class NetworkLevelMatrix:
 
 
 class EnterpriseGraphBuilderTikz:
+    """Converts the list of NetworkLevel into a tikz depiction
+    of the material flow model."""
+
     def __init__(
         self,
         enterprise_name: str,
         list_of_network_level: list[NetworkLevel],
     ) -> None:
+        """
+
+        Args:
[…]
