+++ src/ethos_penalps/network_level.py
@@ -19,6 +19,11 @@ from ethos_penalps.time_data import TimeData
 
 
 class NetworkLevel:
+    """A NetworkLevel is used to simulate one or more chains of subsequent process steps. Each well defined simulation
+    must have at least one NetworkLevel but arbitrary more can be added. Eac networkLevel must contain at least one Sink
+    and one Source.
+    """
+
     def __init__(
         self,
         stream_handler: StreamHandler,
@@ -31,12 +36,24 @@ class NetworkLevel:
         self.load_profile_handler: LoadProfileHandlerSimulation = load_profile_handler
         self.main_source: Source | ProcessChainStorage
         self.main_sink: Sink | ProcessChainStorage
-        self.node_dictionary: dict[ProcessNode] = {}
+        self.node_dictionary: dict[str, ProcessNode] = {}
         self.list_of_process_chains: list[ProcessChain] = []
         self.time_data: TimeData = time_data
-        self.uuid: uuid.UUID = get_new_uuid()
+        self.uuid: str = get_new_uuid()
 
     def create_main_source(self, name: str, commodity: Commodity) -> Source:
+        """Creates a source that marks the start point of the material flow in
+        the NetworkLevel.
+
+        Args:
+            name (str): Name of the source. Is used as key for identification
+                and is displayed in some figures. Must be unique.
+            commodity (Commodity): The commodity that is distributed by the sink.
+
+        Returns:
+            Source: Object that marks the start point of the material flow in
+                the NetworkLevel.
+        """
         source = Source(
             name=name,
             commodity=commodity,
@@ -50,6 +67,19 @@ class NetworkLevel:
     def create_main_sink(
         self, name: str, commodity: Commodity, order_collection: OrderCollection
     ) -> Sink:
+        """Creates a sink that marks the target of the material flow in the NetWorkLevel.
+
+        Args:
+            name (str): Name of the sink. Is displayed in some figures
+                and used as a key for identification. Must be unique.
+            commodity (Commodity): Commodity that is collected by the sink.
+            order_collection (OrderCollection): Object that contains all
+                product order that should be collected by the sink during the
+                simulation.
+
+        Returns:
+            Sink: Marks the target of the material flow in the NetWorkLevel.
+        """
         sink = Sink(
             name=name,
             commodity=commodity,
@@ -67,6 +97,19 @@ class NetworkLevel:
     def create_process_chain_storage_as_source(
         self, name: str, commodity: Commodity
     ) -> ProcessChainStorage:
+        """Creates a ProcessChainStorage and sets it as a Source in
+        this NetworkLevel. The Source determines start start point
+        of the material flow in the NetworkLevel.
+
+        Args:
+            name (str): Name of the ProcessChainStorage. Must be unique.
+            commodity (Commodity): Commodity which is distributed by the
+                Storage.
+
+        Returns:
+            ProcessChainStorage: Object that determines start start point
+                of the material flow in the NetworkLevel
+        """
         process_chain_storage = ProcessChainStorage(
             name=name,
             commodity=commodity,
@@ -84,6 +127,13 @@ class NetworkLevel:
     def add_process_chain_storage_as_sink(
         self, process_chain_storage: ProcessChainStorage
     ):
+        """Adds the ProcessChainStorage instance as a replacement for a sink.
+        Marks the target of the material flow in the NetWorkLevel.
+
+        Args:
+            process_chain_storage (ProcessChainStorage): Object that marks the
+                target of the material flow in this NetworkLevel.
+        """
         process_chain_storage.add_sink_from_next_network_level(
             commodity=process_chain_storage.commodity,
             name=process_chain_storage.name,
@@ -96,7 +146,20 @@ class NetworkLevel:
         )
         self.main_sink = process_chain_storage
 
[…]
