### 2024-03-26 #6358 @julian-belina
@srmnitc @dpzhuX @stoldtjo I have incorporated all requested changes so far.

The API documentation has been improved by using docstrings, which are also displayed in the HTML documentation.  It took a while because the codebase is quite extensive at this point. I will close the corresponding issue. Please leave a comment if you think it needs further improvement. The updated documentation should be available now.

I rewrote the entire paper, and moved the example to my documentation to leave more space for the statement of need section and the other tools and methods section.

@stoldtjo  I would like to address your comments on the paper's structure. I have followed the format suggested by (JOSS)[https://joss.readthedocs.io/en/latest/submitting.html#what-should-my-paper-contain], which differs slightly from other journals.

**Statement of need section**: A detailed description of load profile requirements for energy system models has been provided.  The software's novelty lies not in the implemented simulation paradigm, but rather in its open-source availability. Although previous studies have shared simulation concepts that could model similar systems they did not publish their implementation of these concepts, leaving a reserach gap.

**Other tools and methods** I has been noted that there is a plethora of concepts, simulation studies and commercial tools that could be used for the simulation of manufacturing processes under the consideration of energy aspects. However, there is no ready to use open-source tool that can do this.  There is currently no indication that the source code of previous tools will be published. This tool is open for extension and invites the publication of simulation models in combination with the software.

Additionally, a roadmap to the documentation has been included to showcase the development direction of ETHOS.PeNALPS.

I'm looking forward to your feedback and would like to thank you for your efforts in the review.
