### 2024-03-26 FZJ-IEK3-VSA/ETHOS_PeNALPS#6 @julian-belina
The API documentation has been improved in 1.0.4
