### 2024-03-26 FZJ-IEK3-VSA/ETHOS_PeNALPS#7 @julian-belina
The statement of need has been improved in the 1.0.4 version of the paper.
