### 2024-03-26 FZJ-IEK3-VSA/ETHOS_PeNALPS#8 @julian-belina
The other tools an methods has been improved in the 1.0.4 version of the paper.
