+++ examples/tutorial/_5_connect_four_process_steps/simulation_starter.py
@@ -2,7 +2,7 @@ import datetime
 import logging
 
 from ethos_penalps.data_classes import Commodity, LoadType
-from ethos_penalps.enterprise import Enterprise
+from ethos_penalps.organizational_agents.enterprise import Enterprise
 from ethos_penalps.order_generator import NOrderGenerator
 from ethos_penalps.stream import BatchStreamStaticData, ContinuousStreamStaticData
 from ethos_penalps.time_data import TimeData
