+++ examples/basic_examples/batch_to_batch_1_node_example.py
@@ -5,7 +5,7 @@ from typeguard import install_import_hook
 install_import_hook("ethos_penalps")
 
 from ethos_penalps.data_classes import Commodity, LoadType
-from ethos_penalps.enterprise import Enterprise
+from ethos_penalps.organizational_agents.enterprise import Enterprise
 from ethos_penalps.order_generator import NOrderGenerator
 from ethos_penalps.stream import (
     BatchStreamStaticData,
