+++ src/ethos_penalps/process_nodes/process_node.py
@@ -23,7 +23,7 @@ from ethos_penalps.process_node_communicator import (
     EmptyProductionBranch,
     ProcessNodeCommunicator,
 )
-from ethos_penalps.process_state_handler import ProcessStateHandler
+from ethos_penalps.petri_net.process_state_handler import ProcessStateHandler
 from ethos_penalps.process_step_data import ProcessStepData
 from ethos_penalps.production_plan import ProductionPlan
 from ethos_penalps.simulation_data.simulation_data_branch import (
