+++ examples/tutorial/_1_cooking_example.py
@@ -5,7 +5,7 @@ from typeguard import install_import_hook
 
 install_import_hook("ethos_penalps")
 from ethos_penalps.data_classes import Commodity, LoadType
-from ethos_penalps.enterprise import Enterprise
+from ethos_penalps.organizational_agents.enterprise import Enterprise
 from ethos_penalps.order_generator import NOrderGenerator
 from ethos_penalps.stream import BatchStreamStaticData, ContinuousStreamStaticData
 from ethos_penalps.time_data import TimeData
