+++ test/test_production_plan_post_processor/forming_quenching_and_triming_chain.py
@@ -2,7 +2,7 @@ import datetime
 
 from ethos_penalps.data_classes import Commodity, LoadProfileEntry, LoadType
 from ethos_penalps.utilities.debugging_information import NodeOperationViewer
-from ethos_penalps.enterprise import Enterprise
+from ethos_penalps.organizational_agents.enterprise import Enterprise
 
 from ethos_penalps.post_processing.enterprise_graph_for_failed_run import (
     GraphVisualization,
@@ -10,7 +10,7 @@ from ethos_penalps.post_processing.enterprise_graph_for_failed_run import (
 from ethos_penalps.post_processing.report_generator.process_chain_report_generator import (
     ResultPathGenerator,
 )
-from ethos_penalps.process_chain import ProcessChain
+from ethos_penalps.organizational_agents.process_chain import ProcessChain
 from ethos_penalps.process_nodes.process_chain_storage import ProcessChainStorage
 from ethos_penalps.process_nodes.sink import Sink
 from ethos_penalps.process_nodes.source import Source
