+++ src/ethos_penalps/process_nodes/source.py
@@ -1,48 +1,50 @@
 import datetime
+
 import pandas
-from ethos_penalps.time_data import TimeData
-from ethos_penalps.stream import (
-    BatchStream,
-    ContinuousStream,
-    ContinuousStreamState,
-    BatchStreamState,
-)
 
-from ethos_penalps.stream_handler import StreamHandler
-
-from ethos_penalps.utilities.exceptions_and_warnings import MisconfigurationError
 from ethos_penalps.data_classes import (
-    StaticTimePeriod,
-    OutputBranchIdentifier,
-    TemporalBranchIdentifier,
     Commodity,
+    OrderCollection,
+    OutputBranchIdentifier,
     ProcessChainIdentifier,
+    ProductionOrder,
+    StaticTimePeriod,
+    TemporalBranchIdentifier,
 )
-from ethos_penalps.process_node_communicator import ProcessNodeCommunicator
-from ethos_penalps.process_step_data import ProcessStepData
+from ethos_penalps.load_profile_calculator import LoadProfileHandlerSimulation
 from ethos_penalps.mass_balance import MassBalance
 from ethos_penalps.node_operations import (
-    NodeOperation,
-    DownstreamValidationOrder,
-    UpstreamNewProductionOrder,
-    TerminateProduction,
     DownstreamAdaptionOrder,
+    DownstreamValidationOrder,
+    NodeOperation,
+    TerminateProduction,
     UpstreamAdaptionOrder,
+    UpstreamNewProductionOrder,
 )
-from ethos_penalps.process_node_communicator import EmptyProductionBranch
-from ethos_penalps.utilities.logger_ethos_penalps import PeNALPSLogger
-from ethos_penalps.production_plan import ProductionPlan
-from ethos_penalps.load_profile_calculator import LoadProfileHandlerSimulation
-from ethos_penalps.simulation_data.simulation_data_branch import (
-    IncompleteOutputBranchData,
-    CompleteOutputBranchData,
-    OutputBranchData,
-    IncompleteStreamBranchData,
+from ethos_penalps.process_node_communicator import (
+    EmptyProductionBranch,
+    ProcessNodeCommunicator,
 )
 from ethos_penalps.process_nodes.process_node import ProcessNode
-from ethos_penalps.production_plan import OutputBranchProductionPlan
-from ethos_penalps.data_classes import ProductionOrder, OrderCollection
+from ethos_penalps.process_step_data import ProcessStepData
+from ethos_penalps.production_plan import OutputBranchProductionPlan, ProductionPlan
+from ethos_penalps.simulation_data.simulation_data_branch import (
+    CompleteOutputBranchData,
+    IncompleteOutputBranchData,
+    IncompleteStreamBranchData,
+    OutputBranchData,
+)
 from ethos_penalps.storage import BaseStorage
+from ethos_penalps.stream import (
+    BatchStream,
+    BatchStreamState,
+    ContinuousStream,
+    ContinuousStreamState,
+)
+from ethos_penalps.stream_handler import StreamHandler
+from ethos_penalps.time_data import TimeData
+from ethos_penalps.utilities.exceptions_and_warnings import MisconfigurationError
+from ethos_penalps.utilities.logger_ethos_penalps import PeNALPSLogger
 
 logger = PeNALPSLogger.get_logger_without_handler()
 
