+++ src/ethos_penalps/post_processing/enterprise_graph_for_failed_run.py
@@ -1,22 +1,21 @@
-import webbrowser
-import os
 import logging
+import os
 import warnings
+import webbrowser
 
 import graphviz
 
-from ethos_penalps.utilities.general_functions import ResultPathGenerator
-from ethos_penalps.process_nodes.process_step import ProcessStep
-from ethos_penalps.process_nodes.process_node import ProcessNode
+from ethos_penalps.data_classes import StateConnector
+from ethos_penalps.petri_net.process_state import ProcessState
 from ethos_penalps.process_nodes.process_chain_storage import ProcessChainStorage
+from ethos_penalps.process_nodes.process_node import ProcessNode
+from ethos_penalps.process_nodes.process_step import ProcessStep
 from ethos_penalps.process_nodes.sink import Sink
 from ethos_penalps.process_nodes.source import Source
-from ethos_penalps.utilities.logger_ethos_penalps import PeNALPSLogger
-from ethos_penalps.petri_net.process_state import ProcessState
-from ethos_penalps.stream import ContinuousStream, BatchStream
+from ethos_penalps.stream import BatchStream, ContinuousStream
 from ethos_penalps.stream_handler import StreamHandler
-from ethos_penalps.data_classes import StateConnector
-
+from ethos_penalps.utilities.general_functions import ResultPathGenerator
+from ethos_penalps.utilities.logger_ethos_penalps import PeNALPSLogger
 
 # https://graphviz.readthedocs.io/en/stable/examples.html
 # logging.basicConfig(level=os.environ.get("LOGLEVEL", "DEBUG"))
