+++ src/ethos_penalps/post_processing/report_generator/enterprise_report_generator.py
@@ -11,11 +11,17 @@ import matplotlib.pyplot
 
 from ethos_penalps.data_classes import CurrentProcessNode, LoopCounter
 from ethos_penalps.load_profile_calculator import LoadProfileHandlerSimulation, LoadType
-from ethos_penalps.organizational_agents.network_level import NetworkLevel
 from ethos_penalps.node_operations import ProductionOrder
+from ethos_penalps.organizational_agents.network_level import NetworkLevel
+from ethos_penalps.post_processing.load_profile_entry_post_processor import (
+    LoadProfileEntryPostProcessor,
+)
 from ethos_penalps.post_processing.network_analyzer import (
     NetworkAnalyzer,
     ResultSelector,
+)
+from ethos_penalps.post_processing.post_processed_data_handler import (
+    PostProcessSimulationDataHandler,
 )
 from ethos_penalps.post_processing.process_summary import ProcessOverViewGenerator
 from ethos_penalps.post_processing.report_generator.carpet_plot_page import (
@@ -53,12 +59,6 @@ from ethos_penalps.utilities.debugging_information import (
 )
 from ethos_penalps.utilities.general_functions import ResultPathGenerator
 from ethos_penalps.utilities.logger_ethos_penalps import PeNALPSLogger
-from ethos_penalps.post_processing.load_profile_entry_post_processor import (
-    LoadProfileEntryPostProcessor,
-)
-from ethos_penalps.post_processing.post_processed_data_handler import (
-    PostProcessSimulationDataHandler,
-)
 
 logger = PeNALPSLogger.get_logger_without_handler()
 
