+++ examples/b_pillar_manufacturing/simulation_starter.py
@@ -2,15 +2,14 @@ import datetime
 import logging
 
 from cutting_and_heating_chain import fill_cutting_and_heating_chain
+from ethos_penalps.data_classes import Commodity
+from ethos_penalps.order_generator import NOrderGenerator
+from ethos_penalps.organizational_agents.enterprise import Enterprise
+from ethos_penalps.time_data import TimeData
+from ethos_penalps.utilities.logger_ethos_penalps import PeNALPSLogger
 from forming_quenching_and_trimming_chain import (
     fill_forming_quenching_and_trimming_chain,
 )
-
-from ethos_penalps.data_classes import Commodity
-from ethos_penalps.organizational_agents.enterprise import Enterprise
-from ethos_penalps.order_generator import NOrderGenerator
-from ethos_penalps.time_data import TimeData
-from ethos_penalps.utilities.logger_ethos_penalps import PeNALPSLogger
 
 # Create logger to receiver information about the simulation progress
 logger = PeNALPSLogger.get_human_readable_logger(logging.INFO)
