+++ examples/tutorial/_4_cooking_and_mixer_exclusive_example.py
@@ -2,8 +2,8 @@ import datetime
 import logging
 
 from ethos_penalps.data_classes import Commodity, LoadType
-from ethos_penalps.organizational_agents.enterprise import Enterprise
 from ethos_penalps.order_generator import NOrderGenerator
+from ethos_penalps.organizational_agents.enterprise import Enterprise
 from ethos_penalps.stream import BatchStreamStaticData, ContinuousStreamStaticData
 from ethos_penalps.time_data import TimeData
 from ethos_penalps.utilities.logger_ethos_penalps import PeNALPSLogger

[…]
