+++ test/test_toffee_production/test_toffee_production.py
@@ -1,6 +1,5 @@
 import datetime
 import logging
-
 from test.test_toffee_production.cutting_and_packaging_chain import (
     fill_cutting_and_packaging_chain,
 )
@@ -12,8 +11,8 @@ from test.test_toffee_production.toffee_preparation_chain_2 import (
 )
 
 from ethos_penalps.data_classes import Commodity, LoadType
-from ethos_penalps.organizational_agents.enterprise import Enterprise
 from ethos_penalps.order_generator import NOrderGenerator
+from ethos_penalps.organizational_agents.enterprise import Enterprise
 from ethos_penalps.time_data import TimeData
 from ethos_penalps.utilities.logger_ethos_penalps import PeNALPSLogger
 
