+++ test/production_plan_tests/cutting_and_heating_chain.py
@@ -1,15 +1,14 @@
 import datetime
 
 from ethos_penalps.data_classes import Commodity, LoadProfileEntry, LoadType
-from ethos_penalps.utilities.debugging_information import NodeOperationViewer
 from ethos_penalps.organizational_agents.enterprise import Enterprise
+from ethos_penalps.organizational_agents.process_chain import ProcessChain
 from ethos_penalps.post_processing.enterprise_graph_for_failed_run import (
     GraphVisualization,
 )
 from ethos_penalps.post_processing.report_generator.process_chain_report_generator import (
     ResultPathGenerator,
 )
-from ethos_penalps.organizational_agents.process_chain import ProcessChain
 from ethos_penalps.process_nodes.process_chain_storage import ProcessChainStorage
 from ethos_penalps.process_nodes.sink import Sink
 from ethos_penalps.process_nodes.source import Source
@@ -19,6 +18,7 @@ from ethos_penalps.stream import (
     ContinuousStreamStaticData,
 )
 from ethos_penalps.time_data import TimeData
+from ethos_penalps.utilities.debugging_information import NodeOperationViewer
 from ethos_penalps.utilities.logger_ethos_penalps import PeNALPSLogger
 
 
