### 2024-04-18 #6358 @stoldtjo
Hello @srmnitc and @julian-belina, Please excuse the delay. I reviewed things and I like the results of all the revisions a lot. At this time I only have reason for one minor criticism: The checklist expects a clear statement of need in the docs. I very much like the way this is presented in the paper so the docs should ideally reflect it in a likewise fashion.

With respect to functionality testing I have to confess that I wasn't able to move past included examples. That being said, I trust unit tests, etc. have helped the authors to establish some fundamental and expected functionality.

@srmnitc I suggest you make a call if you find it important that the statement of need is revised in the project docs. Given the fact it's clearly stated in the paper I wouldn't want to hold up the process on this account.
