+++ documentation/ethos_penalps_articles/statement_of_need.md
@@ -1,0 +1,12 @@
+# Statement of Need
+
+Energy system models are tools that provide guidance on future energy systems, which are currently undergoing significant changes to due global efforts to reduce dependence on fossil fuels {cite}`Prina.2020` p. 1. However, building long-term models with high spatial and temporal resolution and transparent input data remains a challenge {cite}`Prina.2020` p. 12. For instance, historical load profiles for the German industrial sector in 2015 are available {cite}`Priesmann.2021` pp. 5–6, while load profiles for other regions are not currently available. Furthermore, decarbonization efforts will cause changes in the industrial sector, creating a need for load profiles of future scenarios. To address the lack of sectoral load profiles for the industry, Boßmann and Stafell {cite}`Bomann.2015` p. 1321 demonstrated the use of a bottom-up approach. Therefore, it is necessary to obtain load profiles of the industrial processes that are part of the industrial sector. However, these profiles are often unavailable for open research due to:
+
+• Companies’ efforts to protect commercial secrets;
+• Missing measurements;
+• Unstructured collection of energy data in companies;
+• Novelty of the industrial processes and their current lack of implementation.
+
+ETHOS.PeNALPS can support the creation of an energy system model by providing load profiles for industrial processes. While many industrial processes and their load profiles have been previously simulated, most have not published load profiles and simulation implementations under an open-source license. This creates a research gap, despite similar work having already been done.
+
+ETHOS.PeNALPS provides modeling capabilities to simulate load profiles of individual production equipment and the logistics between them in a network. Fluctuations of individual production equipment are modeled using a deterministic Petri net of states. The level of detail and temporal resolution of the load profile model depends on the production process features, the level of detail in the process description, and the available input data. To ensure the suitability of a simulated load profile for each energy system model, it is necessary to evaluate its temporal resolution. At lower temporal resolutions, effects may occur that cannot be modeled using a deterministic Petri net of machine states and average energy consumption per state. Furthermore, the temporal resolution of energy system models is constantly evolving. According to {cite}`Prina.2020` p. 10, a temporal resolution of one hour is considered high for long-term energy system models. Currently, studies may require load profiles with a resolution as low as one minute {cite}`Omoyele.2024` pp. 12–13.
