### 2024-04-18 7aa050c
Remove macos 11 from testing due to low calculation capacity on runner
