### 2024-04-18 #6358 @julian-belina
@stoldtjo I copied the statement of need section from my paper to my documentation and fixed the broken link you found. The changes will be released as version 1.0.5. I would like to thank you again for taking the time to review my paper. Especially the literature recommendations were very helpful and guided me in the categorization of my work.

@dpzhuX I also want to thank you for taking the time to review this document. Your feedback helped sharpen the focus and ensure the functionality of the tool.

@srmnitc
I also wanted to thank you for organizing and moderating the review. Assuming you have no further objections, how long will it take to publish?
