### 2024-04-18 FZJ-IEK3-VSA/ETHOS_PeNALPS#2 @julian-belina
Thank you for reporting the bug. It has been fixed in 1.0.5
