### 2024-04-22 #6358 @stoldtjo
> @stoldtjo @dpzhuX thank you both for your detailed reviews once again. Could you please confirm once more that you recommend the publication of this paper in JOSS?

Confirmed. This is certainly worth publishing, also in terms of quality.
