### 2024-04-22 FZJ-IEK3-VSA/ETHOS_PeNALPS#2 @julian-belina
Sorry, for the inconvenience...

To overcome the problematic of dead links, I started to use [linkcheckmd](https://pypi.org/project/linkcheckmd/) for the Readme.md and

`
jupyter-book build documentation/ --builder linkcheck
`
for the documentation. All links should work as expected now.
