### 2024-04-22 #6358 @julian-belina
The latest version number that contains all changes is 1.0.7.
