### 2024-04-22 #6358 @julian-belina
The release is now online and its published on Zenodo
10.5281/zenodo.11032663](https://zenodo.org/doi/10.5281/zenodo.11032663 . Somehow i cant check  the boxes on your last comment @srmnitc
