### 2024-04-22 #6358 @julian-belina
@srmnitc I see there is a problem with at least one of the DOIs. Do i need to create a new release or how can I fix these?
