### 2024-04-23 3bcc202
Update paper references
