### 2024-04-23 #6358 @julian-belina
@srmnitc I checked the references and removed unused sources. Soruces without dois are either llinks to commercial software or e-mail I wrote to certain authors.
