+++ paper/citavi.bib
@@ -267,7 +267,7 @@
 @misc{VolkerWohlgemuth.20.03.2024,
  author = {Wohlgemuth, Volker},
  year = {20.03.2024},
- title = {Software Milan: E-Maill},
+ title = {Software Milan: E-Mail},
  editor = {Belina, Julian}
 }
 
