+++ paper/paper.md
@@ -93,7 +93,7 @@ Stoldt et al. [-@Stoldt.2021] reported the most commonly used simulation tools i
 
 Additionally, Stoldt et al. [-@Stoldt.2021] identified some self-developed standalone tools, but no open-source software was found. Well-documented open-source software and simulation models enhance research reproducibility, leveraging new findings without the need to re-implement state of the art concepts. The licensing of the following software projects has been investigated.
 
-@Wohlgemuth.2006 developed the software "Milan" which is based on a discrete event simulation. According to Wohlgemuth [-@VolkerWohlgemuth.20.03.2024], it was discontinued in 2015. Anderson et al. [-@Andersson.2012b] intended to develop the "EcoProIt tool" for conducting lifecycle assessments based on discrete event simulation. No further information could be found regarding the publication or licensing status of the tool. The "SIMTER tool" was developed in the SIMTER research project [@SallaLind.2009] for combined environmental impact calculations and discrete event simulation. However, information about licensing and distribution is not available. Rippel et al. [-@Rippel.2017] developed the "μ-ProPlAn framework". Rippel [-@Rippel.25.03.2024] stated via e-mail that the software has not been published and is no longer executable due to a lack of maintenance. @Binderbauer.2022 published a software called "Ganymede" that simulates load profiles. However, information about its licensing and distribution is not available.
+@Wohlgemuth.2006 developed the software "Milan" which is based on a discrete event simulation. According to Wohlgemuth (personal communication, 20.03.2024), it was discontinued in 2015. Anderson et al. [-@Andersson.2012b] intended to develop the "EcoProIt tool" for conducting lifecycle assessments based on discrete event simulation. No further information could be found regarding the publication or licensing status of the tool. The "SIMTER tool" was developed in the SIMTER research project [@SallaLind.2009] for combined environmental impact calculations and discrete event simulation. However, information about licensing and distribution is not available. Rippel et al. [-@Rippel.2017] developed the "μ-ProPlAn framework". Rippel (personal communication, 25.03.2024) stated via e-mail that the software has not been published and is no longer executable due to a lack of maintenance. @Binderbauer.2022 published a software called "Ganymede" that simulates load profiles. However, information about its licensing and distribution is not available.
 
 
 
