+++ paper/citavi.bib
@@ -18,7 +18,7 @@
 @misc{AutoModSimulationssoftware.07.11.2019,
  abstract = {Profitieren Sie mit AutoMod von einer dynamischen Simulation und erhalten Sie ein deutlich genaueres Bild des Anlagenverhaltens.},
  author = {{AutoMod Simulationssoftware}},
- year = {21.03.2024},
+ year = {2024},
  title = {AutoMod Simulationssoftware - Simulating Reality},
  url = {https://automod.de/},
  urldate = {21.03.2024}
@@ -56,6 +56,7 @@
  title = {From System Dynamics and Discrete Event to Practical Agent Based Modeling: Reasons, Techniques, Tools},
  booktitle = {International Conference of the System Dynamics Society.},
  year = {2004},
+ url={https://api.semanticscholar.org/CorpusID:6188368},
  file = {systemdyndiscreteeventabmcompared.borshchevfilippov04:Attachments/systemdyndiscreteeventabmcompared.borshchevfilippov04.pdf:application/pdf}
 }
 
@@ -76,7 +77,7 @@
 @misc{LannerGroupLimited.19.03.2024,
  abstract = {WITNESS Process Modeling {\&} Simulation Software from Lanner. Map business processes and workflows to identify opportunities and reduce future risks.},
  editor = {{Lanner Group Limited}},
- year = {19.03.2024},
+ year = {2024},
  title = {WITNESS Simulation Modeling Software},
  url = {https://www.lanner.com/en-gb/technology/witness-simulation-software.html},
  urldate = {19.03.2024}
@@ -86,7 +87,7 @@
 @misc{LLC.19.03.2024,
  abstract = {Simio allows you to create a real-time model of your manufacturing floor that creates that creates a digital twin that can schedule and reschedule instantly!},
  editor = {{Simio LLC}},
- year = {19.03.2024},
+ year = {2024},
  title = {Manufacturing Simulation and Scheduling Software | Simio},
  url = {https://www.simio.com/applications/manufacturing-simulation-software/},
  urldate = {19.03.2024},
@@ -160,19 +161,10 @@
  file = {Proof{\_}154{\_}Rippel:Attachments/Proof{\_}154{\_}Rippel.pdf:application/pdf}
 }
 
-
-@misc{Rippel.25.03.2024,
- author = {Rippel, D.},
- year = {25.03.2024},
- title = {\textgreek{m}-ProPlAn: E-Mail},
- editor = {Belina, Julian}
-}
-
-
 @misc{RockwellAutomation.17.02.2024,
  abstract = {Arena Simulation uses historical data to create a digital twin to analyze system results through Discrete-Event, Flow, and Agent Based modeling methods},
  editor = {{Rockwell Automation}},
- year = {17.02.2024},
+ year = {2024},
  title = {Arena Simulation Software | Arena Simulation Software},
  url = {https://www.rockwellautomation.com/de-de/products/software/arena-simulation.html},
  urldate = {21.03.2024},
@@ -196,7 +188,7 @@
 @misc{SiemensDigitalIndustriesSoftware.15.03.2024,
  abstract = {Use Plant Simulation software for discrete-event simulation and throughput optimization to improve manufacturing performance of your production systems.},
  editor = {{Siemens Digital Industries Software}},
- year = {15.03.2024},
+ year = {2024},
  title = {Plant simulation software | Siemens Software},
  url = {https://plm.sw.siemens.com/en-US/tecnomatix/products/plant-simulation-software/},
  urldate = {19.03.2024},
@@ -233,7 +225,7 @@
 @misc{TheAnyLogicCompany.19.03.2024,
[…]
