### 2024-04-23 #6358 @srmnitc
@julian-belina I opened a pull request now with some small changes in the bib file. Namely fixing the year fields to only year instead of the complete date, and removing email communications from the citations. Please feel free to check again and merge, so simply make the changes yourself.

One other topic is the links: you have a number of links to webpages as citations. You could also convert them to markdown links directly in the text if you prefer this. As it is also fine.
