### 2024-04-23 #6358 @julian-belina
@srmnitc I checked and merged your changes. I would like to keep the links as citations so it remains clear to which webpage they are redirecting. Unfortunately I could not find any permanent references for the commerical software.
