### 2024-05-02 79252e2
Update JOSS  paper
Some minor typographic edits to the JOSS paper
