+++ paper/paper.md
@@ -47,7 +47,7 @@ The user creates the process model based on generic simulation objects. Once the
 
 # Statement of Need
 
-Energy system models are tools that provide guidance on future energy systems, which are currently undergoing significant changes to due global efforts to reduce dependence on fossil fuels. [@Prina.2020 p.1]. However, building long-term models with high spatial and temporal resolution and transparent input data remains a challenge [@Prina.2020 p.12].  For instance, historical load profiles for the German industrial sector in 2015 are available [@Priesmann.2021 p.5-6], while load profiles for other regions are not currently available. Furthermore, decarbonization efforts will cause changes in the industrial sector, creating a need for load profiles of future scenarios. To address the lack of sectoral load profiles for the industry, Boßmann and Stafell [-@Bomann.2015 p.1321] demonstrated the use of a bottom-up approach. Therefore, it is necessary to obtain load profiles of the industrial processes that are part of the industrial sector. However, these profiles are often unavailable for open research due to:
+Energy system models are tools that provide guidance on future energy systems, which are currently undergoing significant changes to due global efforts to reduce dependence on fossil fuels [@Prina.2020 p.1]. However, building long-term models with high spatial and temporal resolution and transparent input data remains a challenge [@Prina.2020 p.12].  For instance, historical load profiles for the German industrial sector in 2015 are available [@Priesmann.2021 p.5-6], while load profiles for other regions are not currently available. Furthermore, decarbonization efforts will cause changes in the industrial sector, creating a need for load profiles of future scenarios. To address the lack of sectoral load profiles for the industry, Boßmann and Stafell [-@Bomann.2015 p.1321] demonstrated the use of a bottom-up approach. Therefore, it is necessary to obtain load profiles of the industrial processes that are part of the industrial sector. However, these profiles are often unavailable for open research due to:
 
 - Companies' efforts to protect commercial secrets;
 - Missing measurements; 
@@ -60,7 +60,7 @@ ETHOS.PeNALPS provides modeling capabilities to simulate load profiles of indivi
 
 # Method
 
-There are four simulation modeling paradigms as shown in \ref{fig:Simulation paradigms}. ETHOS.PeNALPS utilizes an agent-based approach for the nodes of a material flow system. Currently, the most important nodes of the material flow system, the process steps, contain a Petri net to model their activity. The part of the ETHOS.PeNALPS simulation based on the Petri net can be classified as a discrete event simulation. @Borshchev.2004 and @Thiede.2012b p.45-49 provide an introduction and comparison to these paradigms.
+There are four simulation modeling paradigms as shown in Figure \ref{fig:Simulation paradigms}. ETHOS.PeNALPS utilizes an agent-based approach for the nodes of a material flow system. Currently, the most important nodes of the material flow system, the process steps, contain a Petri net to model their activity. The part of the ETHOS.PeNALPS simulation based on the Petri net can be classified as a discrete event simulation. @Borshchev.2004 and @Thiede.2012b provide an introduction and comparison to these paradigms.
 
 ![Simulation paradigms for material flow simulations [@Thiede.2012b p.47] adapted from [@Borshchev.2004 p. 3]. \label{fig:Simulation paradigms}](simulation_methods.png)
 
@@ -85,9 +85,9 @@ The behavior of a process step during request fulfillment is determined by a seq
 
 # Other Tools
 
-There are numerous publications on the simulation of energy features of industrial processes. A collection can be found in [@Stoldt.2019 p.69-73].  However, many of these publications are limited to the presentation of concepts and selected simulation results, without implementation details. This lack of information creates a significant overhead for new research.
[…]
