### 2024-05-02 #6358 @kyleniemeyer
Hi @julian-belina, I made a few minor typographic edits to the paper: https://github.com/FZJ-IEK3-VSA/ETHOS_PeNALPS/pull/10

Can you review and merge these?
