### 2024-05-02 #6358 @julian-belina
@kyleniemeyer
I merged your changes but readded the page number to the Thiede reference.

> Thiede [-@Thiede.2012b p.45-49]
